% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-format.R
\name{add_idf_comment}
\alias{add_idf_comment}
\title{Add IDF Comments}
\usage{
add_idf_comment(val_par, val_com)
}
\arguments{
\item{val_par}{A vector of parameter values.}

\item{val_com}{A vector of corresponding comments.}
}
\value{
A character vector of IDF comments.
}
\description{
This function adds IDF comments to a list of values and corresponding comments.
}
\examples{
add_idf_comment(c("param1", "param2"), c("comment1", "comment2"))
}
\seealso{
Other idf-formatting: 
\code{\link{add_idf_comment_custom}()},
\code{\link{determine_end_of_idf_block}()},
\code{\link{iterate_on_list}()},
\code{\link{iterate_on_output_list}()},
\code{\link{make_idf_block}},
\code{\link{make_idf_block_header}()},
\code{\link{make_idf_block_title}()}
}
\concept{idf-formatting}
