% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{append_ground_temperatures}
\alias{append_ground_temperatures}
\title{Append Ground-Related Temperatures}
\usage{
append_ground_temperatures(temperature_data)
}
\arguments{
\item{temperature_data}{A data frame containing ground temperature data.}
}
\value{
The input data frame with added columns for
FCfactor, Shallow, Deep, and Reflectance.
}
\description{
Calculates and adds ground-related parameters
(FCfactor, Shallow, Deep, Reflectance) to the input data frame.
}
