% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{append_heating_connections}
\alias{append_heating_connections}
\title{Append heating connections to a data frame}
\usage{
append_heating_connections(df, l_loop)
}
\arguments{
\item{df:}{The data frame containing zone information.}

\item{l_loop:}{List of energy system parameters}
}
\value{
A data frame with additional columns:
\code{level}, \code{behind}, \code{ahead}, \code{clg_sp}, \code{htg_schedules},
\code{clg_schedules}, \code{htg_available}, \code{thermostat}, \code{thermostat_ctrl}
}
\description{
This function takes a data frame (\code{df}) and appends columns
indicating the zone behind and ahead of each zone in the data frame.
}
