% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{assign_age_bands}
\alias{assign_age_bands}
\title{Assign age bands}
\usage{
assign_age_bands(age)
}
\arguments{
\item{age}{A numeric vector containing age information.}
}
\value{
A character vector of the same length as \code{age} with age bands.
}
\description{
Assigns age information (\code{age}) into bands: "child", "young",
or "adult".
}
