% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{assign_appliances_profile}
\alias{assign_appliances_profile}
\title{Assigns Appliances to Zones}
\usage{
assign_appliances_profile(l_sched, df_zones)
}
\arguments{
\item{l_sched}{A list containing appliance data.}

\item{df_zones}{A data frame defining zones and their characteristics.}
}
\value{
A data frame with appliances and assigned zones.
}
\description{
Assigns appliances to specific zones based on their type.
}
