% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{assign_comfort_ranges}
\alias{assign_comfort_ranges}
\title{Assign Comfort Ranges}
\usage{
assign_comfort_ranges(oscillation_value)
}
\arguments{
\item{oscillation_value}{The oscillation value.}
}
\value{
The corresponding comfort range.
}
\description{
Determines the comfort range based on the given oscillation value.
}
