% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{assign_custom_inputs}
\alias{assign_custom_inputs}
\title{Assigns custom inputs to model parameters}
\usage{
assign_custom_inputs(l_i, l_defs, s_main)
}
\arguments{
\item{l_i}{List of custom input values.}

\item{l_defs}{List of model default parameters.}

\item{s_main}{Model data structure.}
}
\value{
A list containing the modified \code{s_main}, \code{l_defs}, and original \code{l_i} inputs.
}
\description{
This function assigns custom input values to model parameters within the
provided \code{l_defs} and \code{s_main} objects based on the input list \code{l_i}.
}
