% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{assign_devices}
\alias{assign_devices}
\title{Assign device information based on loop type}
\usage{
assign_devices(i, d_lp, d_it)
}
\arguments{
\item{i:}{Character string representing the loop type}

\item{d_lp:}{Data frame containing loop information (groups, types, ...)}

\item{d_it:}{Data frame containing system information}
}
\value{
A list containing information about plant inlet/outlet,
demand inlet/outlet, branches, and connectors for the loop type.
}
\description{
This function extracts information about nodes-connections
for a specific loop type (\code{i}) from a loop data frame (\code{d_lp}).
}
