% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_door_type}
\alias{assign_door_type}
\title{Assigns door types and properties}
\usage{
assign_door_type(dfa, dfb, df_hlc)
}
\arguments{
\item{dfa}{A data frame containing door U-value information.}

\item{dfb}{A data frame containing door count information.}

\item{df_hlc}{A data frame containing construction and material data.}
}
\value{
A list containing:
\itemize{
\item idf: A character vector representing the IDF blocks for the doors.
\item summary: A list containing data frames for exterior and interior door summaries.
}
}
\description{
Creates IDF blocks for exterior and interior doors based on provided data.
}
