% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{assign_font}
\alias{assign_font}
\title{Assigns Font Name for Plotting}
\usage{
assign_font()
}
\value{
The specified font name.
}
\description{
Assigns the font name for plotting, checking availability on macOS.
}
\details{
This function assigns the font name for plotting, as defined
in \code{the$.enhub.config$font$name}. On macOS, it checks the availability
of the font using \code{fc-list} command. If the font is available, a
message is printed; otherwise, a warning is issued.
}
