% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{assign_house_layout}
\alias{assign_house_layout}
\title{Assign Layout Templates Based on House Specifications}
\usage{
assign_house_layout(
  k_sto,
  k_rom,
  k_bed,
  k_att,
  k_bsm,
  k_dim = c(4, 4),
  k_shp = "R",
  l_lay = t_layouts
)
}
\arguments{
\item{k_sto}{Integer. Number of storeys in the house.}

\item{k_rom}{Integer. Total number of rooms in the house.}

\item{k_bed}{Character. The lowest storey with a bedroom. Possible values: 'GF' (Ground Floor), '1F' (First Floor), etc.}

\item{k_att}{Logical. Indicates whether the house has an attic (TRUE) or not (FALSE).}

\item{k_bsm}{Logical. Indicates whether the house has a basement (TRUE) or not (FALSE).}

\item{k_dim}{Numeric vector of length 2. Dimensions of the layout, default is c(4,4).}

\item{k_shp}{Character. Shape of the layout. Default is 'R' for Rectangular. Other shapes are not yet implemented.}

\item{l_lay}{List. A list of layout templates.}
}
\value{
A vector of layout templates corresponding to the specified house configuration.
}
\description{
This function assigns layout templates for a house based on the number of rooms,
number of storeys, the presence of an attic and basement, and the shape and dimensions of the layout.
}
\examples{
\dontrun{
assign_house_layout(k_sto=2, k_rom=5, k_bed='1F', k_att=TRUE, k_bsm=TRUE, k_dim=c(4,4), k_shp='R', l_lay=t_layouts)
assign_house_layout(k_sto=1, k_rom=3, k_bed='GF', k_att=FALSE, k_bsm=FALSE, k_dim=c(2,2), k_shp='R', l_lay=t_layouts)
}
}
