% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_internal_floor}
\alias{assign_internal_floor}
\title{Assigns internal floor properties}
\usage{
assign_internal_floor(idfTo, idf.label = "Interior Floor")
}
\arguments{
\item{idfTo}{A character vector representing the existing IDF content.}

\item{idf.label}{A character string specifying the label for the internal floor (default: "Interior Floor").}
}
\value{
A character vector representing the updated IDF content with an internal floor definition.
}
\description{
Inserts an "Interior Floor" definition within existing IDF content.
}
