% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appliances.R
\name{assign_loads}
\alias{assign_loads}
\title{Assigns loads to IDF objects based on schedule data}
\usage{
assign_loads(d_app, d_sch, l_sup = l_idf_blocks)
}
\arguments{
\item{d_app}{Data frame containing appliance data.}

\item{d_sch}{Data frame containing occupancy schedules.}

\item{l_sup}{List of IDF block objects (default: l_idf_blocks).}
}
\value{
List of modified IDF block objects.
}
\description{
This function assigns loads from an appliance data frame to IDF objects based
on occupancy schedules and fuel types.
}
