% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_loft_material}
\alias{assign_loft_material}
\title{Assigns properties for loft insulation material}
\usage{
assign_loft_material(idfTo, l_IR, varThk = 270)
}
\arguments{
\item{idfTo}{A character vector representing the existing IDF content.}

\item{l_IR}{A list representing the surface object for the loft ceiling.}

\item{varThk}{A numeric value specifying the insulation thickness in millimeters (default: 270).}
}
\value{
A list containing:
\itemize{
\item summary: An updated surface object (l_IR) with insulation properties.
\item idf: A character vector representing the updated IDF content for the loft.
}
}
\description{
Updates a surface object (l_IR) with insulation properties and creates IDF content for the loft construction.
}
