% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{assign_occupants_profile}
\alias{assign_occupants_profile}
\title{Assigns Occupancy Profiles to Zones}
\usage{
assign_occupants_profile(df_presence, df_zones)
}
\arguments{
\item{df_presence}{A data frame containing presence data across hours.}

\item{df_zones}{A data frame defining zones and their characteristics.}
}
\value{
A data frame with presence status for each zone and hour.
}
\description{
Assigns occupancy profiles to each zone based on overall presence data.
}
