% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{assign_presence_by_people}
\alias{assign_presence_by_people}
\title{Assign presence to activity zones for a specific person}
\usage{
assign_presence_by_people(i, d_in, d_ref)
}
\arguments{
\item{i}{The person identifier.}

\item{d_in}{A data frame containing presence probabilities for the person.}

\item{d_ref}{A data frame containing reference data on activity zones.}
}
\value{
A data frame with presence probabilities for each zone across time.
}
\description{
This function assigns presence probabilities to activity zones
based on a reference data frame and a person identifier.
}
