% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{assign_room_layout}
\alias{assign_room_layout}
\title{Iteratively assign room layout and matching components}
\usage{
assign_room_layout(
  idm,
  dim_mtx,
  k_layout,
  k_levels,
  k_attach,
  d_room = d_rooms,
  d_layout = t_layouts
)
}
\arguments{
\item{idm}{selected storey number}

\item{dim_mtx}{matrix dimensions for internal layout}

\item{k_layout}{reference of available storeys/floors in building}

\item{k_levels}{reference of available storeys/floors in building}

\item{k_attach}{reference of attachment condition of building}

\item{d_room}{reference for available room uses}

\item{d_layout}{reference for available layouts templates}
}
\value{
a dictionary of room types per floor
}
\description{
Iteratively assign room layout and matching components
}
