% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{assign_terrain}
\alias{assign_terrain}
\title{Assign Terrain Type}
\usage{
assign_terrain(urban_classification, sparse_classification)
}
\arguments{
\item{urban_classification}{A character string representing the urban classification.}

\item{sparse_classification}{A character string representing the sparse classification.}
}
\value{
A character string representing the assigned terrain type.
}
\description{
This function determines the terrain type based on the provided urban and sparse classifications.
It maps urban classifications to general terrain types and then overrides the result to 'Ocean'
if the sparse classification is 'sparse'.
}
