% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_top_block_insulation}
\alias{assign_top_block_insulation}
\title{Assigns properties for top block insulation material}
\usage{
assign_top_block_insulation(idfTo, l_ER, varThk = 270)
}
\arguments{
\item{idfTo}{A character vector representing the existing IDF content.}

\item{l_ER}{A list representing the surface object for the top block.}

\item{varThk}{A numeric value specifying the insulation thickness in millimeters (default: 270).}
}
\value{
A list containing:
\itemize{
\item summary: An updated surface object (l_ER) with insulation properties.
\item idf: A character vector representing the updated IDF content for the top block.
}
}
\description{
Updates a surface object (l_ER) with insulation properties and creates IDF content for the top block construction.
}
