% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{assign_vertical_voids}
\alias{assign_vertical_voids}
\title{Assign rooms with vertical inter-connections or voids.
Currently only 'Stairwell' zones are considered.}
\usage{
assign_vertical_voids(l_romly)
}
\arguments{
\item{l_romly}{list of layout parameters}
}
\value{
updated list of layout parameters
}
\description{
Assign rooms with vertical inter-connections or voids.
Currently only 'Stairwell' zones are considered.
}
