% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_void_boundary}
\alias{assign_void_boundary}
\title{Assigns a void air boundary construction}
\usage{
assign_void_boundary(l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup}{A list of IDF blocks, used as a base for the new construction.
(Default: \code{l_idf_blocks})}
}
\value{
A character vector representing the IDF block for the void air boundary.
}
\description{
Creates an IDF block for a void air boundary construction with specified parameters.
}
