% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assign_window_type}
\alias{assign_window_type}
\title{Assigns window type and properties}
\usage{
assign_window_type(dfa, dfb, df_hlc, l_sup = l_idf_blocks)
}
\arguments{
\item{dfa}{A data frame containing window type and frame information.}

\item{dfb}{A data frame containing additional window parameters.}

\item{df_hlc}{A data frame containing construction and material data.}

\item{l_sup}{A list of IDF blocks, used as a base for the new constructions.}
}
\value{
A list containing:
\itemize{
\item idf: A character vector representing the IDF blocks for the windows.
\item summary: A data frame summarizing the window properties.
}
}
\description{
Creates IDF blocks for window types, including exterior and interior windows,
based on provided data.
}
