% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{assume_heated_underfloor}
\alias{assume_heated_underfloor}
\title{Checks if heated underfloor system is assumed}
\usage{
assume_heated_underfloor(hc)
}
\arguments{
\item{hc}{A character string representing the energy system code.}
}
\value{
A logical value indicating whether a heated underfloor system is assumed (TRUE) or not (FALSE).
}
\description{
Determines if a heated underfloor system is assumed based on the provided energy system code.
}
