% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{check_enhub_setup}
\alias{check_enhub_setup}
\title{Checks EnHub Setup}
\usage{
check_enhub_setup()
}
\description{
Verifies the existence of required directories for EnHub operations.
}
\details{
This function checks if the specified EnergyPlus and gzip directories exist.
If either directory is missing, an error is thrown.
}
