% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialiser.R
\name{check_external_dependencies}
\alias{check_external_dependencies}
\title{Check external dependencies}
\usage{
check_external_dependencies(
  fileout = "config/external.yaml",
  dirbase = getwd(),
  redefine = FALSE
)
}
\arguments{
\item{fileout}{The path to the YAML file where the configuration will
be saved. Default is \code{config/external.yaml}.}

\item{dirbase}{The path of the directory where fileout is to be stored.
Default is \code{getwd()}.}

\item{redefine}{A logical value indicating whether the configuration is
working for the user or it needs to be generated. Default is \code{FALSE},
although it checks if the file exists.}
}
\value{
None
}
\description{
This function checks the location of \code{EnergyPlus} and \code{Gzip}
based on the operating system. It prompts the user to confirm or edit
the configuration values and saves the updated configuration to a YAML
file located in \verb{<root>/config/external.yaml}.

Typical content of the YAML file will be:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{eplus:
 dir: /Applications/EnergyPlus-24-1-0/
gzip:
 bin: /usr/bin/gunzip
font:
 name: CMU Concrete
}\if{html}{\out{</div>}}
}
