% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typology.R
\name{combine_ehs}
\alias{combine_ehs}
\title{Combines EHS Data Sets into a Single Data Frame}
\usage{
combine_ehs(l_set = s_ehs_2011_ext)
}
\arguments{
\item{l_set}{A list containing EHS data sets (e.g., dwelling information, ventilation, etc.).}
}
\value{
A data frame containing the combined EHS data.
}
\description{
This function merges various EHS data sets (assumed to be part of a larger list) into a single data frame.
It also performs some data cleaning and transformation steps.
}
