% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{compile_envelope_heat_loss_data}
\alias{compile_envelope_heat_loss_data}
\title{Collect and clean building element data}
\usage{
compile_envelope_heat_loss_data(l_flr, l_glz, l_opq, l_int, l_bld)
}
\arguments{
\item{l_flr}{Data table containing floor information}

\item{l_glz}{Data table containing glazing information}

\item{l_opq}{Data table containing opaque element (wall etc.) information}

\item{l_int}{Data table containing internal element information}

\item{l_bld}{Data table containing building information}
}
\value{
List containing cleaned data tables for floors, openings,
opaques, internals, and building.
}
\description{
This function takes a list of data tables containing
information about different building elements (floors, glazing,
opaques, internals, and building) and performs basic cleaning.
}
