% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{compose_geometry_idf}
\alias{compose_geometry_idf}
\title{Generate and concatenate multiple sections required for the generation
of the geometry of a house in idf format}
\usage{
compose_geometry_idf(l_zns, l_blc, l_glz, l_drr)
}
\arguments{
\item{l_zns}{list of zones, incl. vertices and pre-generated idf}

\item{l_blc}{list of opaque surfaces, incl. vertices and pre-generated idf}

\item{l_glz}{list of glazing, incl. vertices and pre-generated idf}

\item{l_drr}{list of doors, incl. vertices and pre-generated idf}
}
\value{
a sequence of strings in idf format
}
\description{
Generate and concatenate multiple sections required for the generation
of the geometry of a house in idf format
}
\examples{
\dontrun{
compose_geometry_idf(l_zones, l_blocks, l_glazing, l_doors)
}
}
