% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus-wrapper.R
\name{convert_idf_to_json}
\alias{convert_idf_to_json}
\title{Convert IDF to EP-JSON}
\usage{
convert_idf_to_json(filepath)
}
\arguments{
\item{filepath}{The path to the input IDF file.}
}
\value{
None. The function performs file conversion and renaming operations.
}
\description{
A wrapper that converts an IDF file to JSON format using
EnergyPlus the argument \code{--convert-only} for the binary \verb{/energyplus}

For example, the following IDF object:

\if{html}{\out{<div class="sourceCode fortran">}}\preformatted{!-   =========== ALL OBJECTS IN CLASS: BOILER:HOTWATER ===========

Boiler:HotWater,        !- 
   hw-standard-boiler, !- Name
   NaturalGas,         !- Fuel Type
   Autosize,           !- Nominal Capacity \{W\}
   0.845,              !- Nominal Thermal Efficiency
   LeavingBoiler,      !- Efficiency Curve Temperature Evaluation Variable
   Boiler Efficiency,  !- Normalized Boiler Efficiency Curve Name
   Autosize,           !- Design Water Flow Rate \{m3/s\}
   0,                  !- Minimum Part Load Ratio
   1,                  !- Maximum Part Load Ratio
   1,                  !- Optimum Part Load Ratio
   n-11005,            !- Boiler Water Inlet Node Name
   n-11006,            !- Boiler Water Outlet Node Name
   99,                 !- Water Outlet Upper Temperature Limit \{C\}
   NotModulated,       !- Boiler Flow Mode
   0,                  !- On Cycle Parasitic Electric Load \{W\}
   1;                  !- Sizing Factor
}\if{html}{\out{</div>}}

is converted to the following JSON object:

\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
   "Boiler:HotWater": \{
       "hw-standard-boiler": \{
           "boiler_flow_mode": "NotModulated",
           "boiler_water_inlet_node_name": "n-11005",
           "boiler_water_outlet_node_name": "n-11006",
           "design_water_flow_rate": "Autosize",
           "efficiency_curve_temperature_evaluation_variable": "LeavingBoiler",
           "fuel_type": "NaturalGas",
           "maximum_part_load_ratio": 1,
           "minimum_part_load_ratio": 0,
           "nominal_capacity": "Autosize",
           "nominal_thermal_efficiency": 0.845,
           "normalized_boiler_efficiency_curve_name": "Boiler Efficiency",
           "on_cycle_parasitic_electric_load": 0,
           "optimum_part_load_ratio": 1,
           "sizing_factor": 1,
           "water_outlet_upper_temperature_limit": 99
       \}
 \}
}\if{html}{\out{</div>}}
}
\seealso{
Other EnergyPlus evaluation: 
\code{\link{extract_simulation_outputs}()},
\code{\link{generate_idf}()},
\code{\link{generate_idf_multiple}()},
\code{\link{read_idf_outcomes}()},
\code{\link{read_input_file}()},
\code{\link{read_input_summary}()},
\code{\link{run_energy_plus}()},
\code{\link{simulate_available_idfs}()}
}
\concept{EnergyPlus evaluation}
