% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{convert_mtoe_to_kwh}
\alias{convert_mtoe_to_kwh}
\title{Converts Energy Units Between Mtoe and kWh}
\usage{
convert_mtoe_to_kwh(data, unit_in = "kWh")
}
\arguments{
\item{data}{The energy values to be converted.}

\item{unit_in}{The input unit of the data ("kWh" or "mtoe").}
}
\value{
A numeric vector of converted energy values.
}
\description{
Converts energy values between million tonnes of oil
equivalent (Mtoe) and kilowatt-hours (kWh) based on
the conversion factor of 1 Mtoe = 11630 kWh (ref: DUKES).
}
\examples{
\dontrun{
convert_mtoe_to_kwh(1, "mtoe")  # Convert 1 Mtoe to kWh
}
}
