% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--layouts.R
\docType{data}
\name{d_rooms}
\alias{d_rooms}
\title{Data with room information}
\format{
A data frame with 51 rows and 10 columns:
\describe{
\item{id}{\code{<int>}: Unique identifier for each room layout.}
\item{code}{\code{<chr>}: Code representing the room layout (e.g., "ZN000", "ZN001").}
\item{name}{\code{<chr>}: Name of the room layout (e.g., "Living Room / Drawing", "Parlour").}
\item{colour}{\code{<chr>}: Hex color code representing the color associated with the room layout (e.g., "#000000", "#5bb08a").}
\item{description}{\code{<chr>}: Description of the room layout (e.g., "\if{html}{\out{<void>}} used to represent complex geometries.", "A room for relaxing and socializing.").}
\item{ehs.function}{\code{<chr>}: Function of the room layout in the context of energy and heating systems (e.g., "living").}
\item{setpoint.temp}{\code{<int>}: Setpoint temperature for the room layout in degrees Celsius (e.g., 21, 18).}
\item{cat.A}{\code{<dbl>}: Category A rating for the room layout (e.g., 1.5, 2.0).}
\item{cat.B}{\code{<dbl>}: Category B rating for the room layout (e.g., 1.0, 1.5).}
\item{cat.C}{\code{<dbl>}: Category C rating for the room layout (e.g., 0.5, 1.5).}
}
}
\usage{
d_rooms
}
\description{
This data frame contains information about various room layouts, including their IDs, codes, names, colours, descriptions, functions, setpoint temperatures, and category ratings.
}
\examples{
\dontrun{
# Displaying the whole list
d_rooms
}
}
\seealso{
Other room layouts: 
\code{\link{t_layouts}}
}
\concept{room layouts}
\keyword{datasets}
