% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--weather.R
\docType{data}
\name{d_stations}
\alias{d_stations}
\title{Data with weather station information}
\format{
A data frame containing weather station information.
\itemize{
\item station: An integer representing the station ID.
\item place: A character string representing the name of the place where the station is located.
\item admin: A character string representing the administrative region code (e.g., "ENG" for England).
\item country: A character string representing the country code (e.g., "GBR" for Great Britain).
\item lat: A numeric value representing the latitude of the station.
\item lon: A numeric value representing the longitude of the station.
\item tzn: An integer representing the time zone offset from UTC.
\item elv: An integer representing the elevation of the station in meters.
\item hist: A logical value indicating whether historical data is available for the station.
\item filepack: A character string representing the name of the file package containing the data.
\item uuid: A character string representing the unique identifier for the station.
\item regional_default: A character string representing the regional default settings (if any).
}
}
\usage{
d_stations
}
\description{
This object contains the summary table of weather station information.
}
\examples{
\dontrun{
# Displaying the whole list
d_stations
}
}
\keyword{datasets}
