% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{decode_cuboid_inputs}
\alias{decode_cuboid_inputs}
\title{Parse defined cuboid information}
\usage{
decode_cuboid_inputs(lc, ld)
}
\arguments{
\item{lc}{list of custom cuboid inputs}

\item{ld}{list of default inputs}
}
\value{
list of parsed attributes
}
\description{
Parse defined cuboid information
}
