% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{define_emitters}
\alias{define_emitters}
\title{Define emitters based on input data}
\usage{
define_emitters(di, dz, ln)
}
\arguments{
\item{di}{A data frame containing input data.}

\item{dz}{A data frame containing zone data.}

\item{ln}{A list of data frames.}
}
\value{
A data frame containing emitter information.
}
\description{
This function defines emitter information based
on provided data.
}
