% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{define_heat_pump_components}
\alias{define_heat_pump_components}
\title{Create data frame of HP components with corresponding
inlets and outlets}
\usage{
define_heat_pump_components(ln, di)
}
\arguments{
\item{ln:}{List containing main device properties}

\item{di:}{Data frame containing system information}
}
\value{
A data frame indicating connections to heat pump
}
\description{
This function generates a data frame indicating connections
to heat pump, using an iterator (\code{ln}) defining main device properties,
and a data frame (di) with system information.
}
