% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{define_heating_components}
\alias{define_heating_components}
\title{Define heating fuels and heater types based on system codes}
\usage{
define_heating_components(k_msh, k_dhw)
}
\arguments{
\item{k_msh:}{Integer representing the master system type code.}

\item{k_dhw:}{Integer representing the domestic hot water code.}
}
\value{
A list with four elements:
\itemize{
\item \code{main_fuel}: Character vector indicating the main space-heating fuel
\item \code{main_heater}: Character vector indicating the main space-heating heater
\item \code{main_device}: Character vector indicating the main space-heating device
\item \code{main_emitter}: Character vector indicating the main space-heating emitter
\item \code{main_efficiency}: Numeric vector indicating the main space-heating efficiency
\item \code{main_cylinder}: Numeric vector indicating the main cylinder volume
\item \code{main_regime}: Character vector indicating the main space-heating
\item \code{main_emitters}: Character vector indicating the main space-heating availability
\item \code{dhw_fuel}: Character vector indicating the domestic hot-water fuel
\item \code{dhw_heater}: Character vector indicating the domestic hot-water heater
\item \code{dhw_device}: Character vector indicating the domestic hot-water device
\item \code{radiators_temperature}: Character vector indicating the radiators design temperature
\item \code{curves}: Character vector indicating the curves associated to the energy system
}
}
\description{
This function takes master system type (\code{k_msh}) and
domestic hot water type (\code{k_dhw}) codes and determines the associated
heating fuel and heater type.
}
\examples{
\dontrun{
k_msh=l_loop$heating_codes$msh
k_dhw=l_loop$heating_codes$dhw
define_heating_components(k_msh, k_dhw)
}
}
