% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{define_loop_parameters}
\alias{define_loop_parameters}
\title{Define loop parameters based on loop type}
\usage{
define_loop_parameters(k_loop)
}
\arguments{
\item{k_loop:}{Character string representing the loop type code.}
}
\value{
A list containing loop parameters for the specified type.
The list can contain additional loop types if defined by \code{k_loop}.
}
\description{
This function takes a loop type code (\code{k_loop}) and
defines a list of parameters for that loop.
}
\examples{
\dontrun{
define_loop_parameters("zone-loop-double")
}
}
