% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{define_selected_system_schedule}
\alias{define_selected_system_schedule}
\title{Define selected system schedule}
\usage{
define_selected_system_schedule(i_code, l_loop, l_sys, l_scd)
}
\arguments{
\item{i_code}{The energy system code.}

\item{l_loop}{A list containing loop parameters.}

\item{l_sys}{A list of loop parameters}

\item{l_scd}{A list of generated schedules}
}
\value{
A list containing the defined schedule.
}
\description{
This function defines a selected system schedule based on
the provided energy system code and loop data.
}
