% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{define_weather_file_source}
\alias{define_weather_file_source}
\title{Define Weather File Source}
\usage{
define_weather_file_source(hid, lstEHS = s_ehs_2011_ext)
}
\arguments{
\item{hid}{The ID of the dwelling.}

\item{lstEHS}{A list containing EHS data.}
}
\value{
A list containing the weather file source and region information.
}
\description{
This function defines the weather file source for a given dwelling.
It determines the weather file source based on dwelling location and user preferences.
}
