% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-format.R
\name{determine_end_of_idf_block}
\alias{determine_end_of_idf_block}
\title{Determine End of IDF Block}
\usage{
determine_end_of_idf_block(ini, txt)
}
\arguments{
\item{ini}{The starting index of the IDF block.}

\item{txt}{The text to search.}
}
\value{
A list containing the start index, end index, and block label.
}
\description{
This function determines the end of an IDF (Input Data File) block.
}
\examples{
\dontrun{
determine_end_of_idf_block(1, c("line1", "line2", ";"))
}
}
\seealso{
Other idf-formatting: 
\code{\link{add_idf_comment}()},
\code{\link{add_idf_comment_custom}()},
\code{\link{iterate_on_list}()},
\code{\link{iterate_on_output_list}()},
\code{\link{make_idf_block}},
\code{\link{make_idf_block_header}()},
\code{\link{make_idf_block_title}()}
}
\concept{idf-formatting}
