% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{display_chronometer}
\alias{display_chronometer}
\title{Displays Elapsed Time}
\usage{
display_chronometer(proc, pt)
}
\arguments{
\item{proc}{A \code{proc.time()} object representing the end time.}

\item{pt}{A \code{proc.time()} object representing the start time (defaults
to the global variable \code{.varChrono}).}
}
\description{
Calculates and displays the elapsed time
from \code{proc.time()} values.
}
\details{
The function calculates the difference between the \code{proc}
and \code{pt} \code{proc.time()} objects. It replaces any NA values with zero
and then displays the sum of the difference in minutes, rounded
to two decimal places.
}
