% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{display_epw_info}
\alias{display_epw_info}
\title{Display Energy Plus Weather (EPW) Info}
\usage{
display_epw_info(l_info)
}
\arguments{
\item{l_info}{A list containing the EPW information.}
}
\value{
The function prints the formatted JSON information
of the EPW file.
}
\description{
This function checks if the EPW file exists and,
if so, converts the EPW info to JSON format and prints it.
}
