% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{edit_surface_width}
\alias{edit_surface_width}
\title{Adjusts surface width based on target value}
\usage{
edit_surface_width(d_source, sel)
}
\arguments{
\item{d_source}{A numeric vector representing the original surface widths.}

\item{sel}{The target total width.}
}
\value{
A numeric vector of adjusted surface widths, rounded to four decimal places.
}
\description{
This function adjusts the values in \code{d_source} proportionally to match the total \code{sel}.
}
