% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{edit_thermal_properties}
\alias{edit_thermal_properties}
\title{Adjusts thermal properties to achieve target U-value}
\usage{
edit_thermal_properties(d_thr, u_val)
}
\arguments{
\item{d_thr}{A data frame containing thermal properties.
Expected columns: \code{conductivity__w_m_k_}, \code{thickness__m_}.}

\item{u_val}{The desired U-value (W/m^2*K).}
}
\value{
The modified data frame with adjusted thermal conductivity.
}
\description{
This function modifies the thermal conductivity of materials in a dataset
to achieve a specified U-value while maintaining the same thermal resistance.
}
