% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{encode_renewable}
\alias{encode_renewable}
\title{Encodes renewable energy information.}
\usage{
encode_renewable(id_ren)
}
\arguments{
\item{id_ren}{A numeric or character string representing renewable energy types.
If numeric, it's interpreted as a binary code. If character, it's a string of binary digits.}
}
\value{
A logical vector indicating the presence of different renewable energy types:
w.biogas, x.turbine, y.water, z.solarpv.
}
\description{
Converts a numeric or character string representing renewable energy types into a logical vector.
}
