% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhub-package.R
\docType{package}
\name{enhub-package}
\alias{enhub}
\alias{enhub-package}
\title{enhub: A platform for the dynamic simulation of national housing stocks}
\description{
To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "enhub")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EnHub-UK/enhub.git}
}

}
\author{
\strong{Maintainer}: Gustavo Sousa \email{g.sousa@sheffield.ac.uk} (\href{https://orcid.org/0000-0001-7433-874X}{ORCID})

}
\keyword{internal}
