% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{estimate_heat_loss_coefficient}
\alias{estimate_heat_loss_coefficient}
\title{Estimates heat loss coefficient for a building}
\usage{
estimate_heat_loss_coefficient(
  l_zn,
  l_mt,
  l_te,
  l_ar,
  d_dy,
  i_odt = NULL,
  i_gnt = NULL,
  i_hdd = NULL
)
}
\arguments{
\item{l_zn}{A list containing zone information (volume, etc.).}

\item{l_mt}{A list containing material information (U-values, etc.).}

\item{l_te}{A list containing temperature data (outdoor, setpoints).}

\item{l_ar}{A list containing air-related data (ACH, etc.).}

\item{d_dy}{A data frame containing drybulb temperature data (optional).}

\item{i_odt}{A numeric value representing outdoor design temperature (optional, defaults to coldest outdoor temperature).}

\item{i_gnt}{A numeric value representing ground temperature (optional).}

\item{i_hdd}{A numeric value representing heating degree days (optional, defaults to total heating degree days).}
}
\value{
A list containing heat loss details for various components.
}
\description{
This function calculates the heat loss coefficient (HLC) for a building
based on various building characteristics and weather data. It also estimates
heat loss due to fabric conduction and ventilation.

The heat loss coefficient (\eqn{HLC}) and heat capacity (\eqn{Q_{h}}) are calculated as follows:

\deqn{ Qf/ΔT = \sum UxA }

\deqn{ Qv/ΔT = 0.33 × n × V}

  \deqn{ HLC = (Q_{f} + Q_{v})/ΔT}

\deqn{ Q_{h} = T_{in-base} × HLC}
}
