% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{estimate_heat_loss_parameter}
\alias{estimate_heat_loss_parameter}
\title{Estimates heat loss parameters for a given housing ID}
\usage{
estimate_heat_loss_parameter(
  hid,
  l_param,
  d_ids = the$core_index,
  l_ehs = s_ehs_2011_ext
)
}
\arguments{
\item{hid}{Housing ID}

\item{l_param}{List of parameters (unused in this function)}

\item{d_ids}{Data frame containing ID information (default: the$core_index)}

\item{l_ehs}{List of energy and housing survey data (default: s_ehs_2011_ext)}
}
\value{
List containing heat loss parameters for the given housing ID
}
\description{
Calculates heat loss parameters for a specific
housing ID based on provided data.
}
