% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{estimate_infiltration_rates}
\alias{estimate_infiltration_rates}
\title{Estimates infiltration rates}
\usage{
estimate_infiltration_rates(d_ref, d_age, l_dim, i_ven = TRUE)
}
\arguments{
\item{d_ref}{Reference data.}

\item{d_age}{Building age data.}

\item{l_dim}{Building dimensions data.}

\item{i_ven}{Logical flag indicating whether to include ventilation.}
}
\value{
A data frame containing estimated infiltration rates.
}
\description{
Calculates infiltration rates based on building age, room dimensions, and ventilation flag.
}
