% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{estimate_solar_parameters}
\alias{estimate_solar_parameters}
\title{Estimates solar parameters for a specific dwelling.}
\usage{
estimate_solar_parameters(
  idc,
  l_EPW,
  l_EHS = s_ehs_2011_ext,
  l_ONS = s_ons,
  l_SAP = s_sap,
  l_BREDEM = s_bredem,
  l_NOABL = s_noabl,
  l_MetO = s_metoffice
)
}
\arguments{
\item{idc}{Character string representing the dwelling code.}

\item{l_EPW}{A list containing weather data (assumed to include latitude information).}

\item{l_EHS}{A list containing dwelling and household information (assumed to include dwelling information and complementary data).}

\item{l_ONS}{(Optional) A list containing additional data (not used in this function).}

\item{l_SAP}{A list containing solar access pattern data.}

\item{l_BREDEM}{A list containing horizontal radiation data.}

\item{l_NOABL}{A list containing wind data (assumed to include regional data).}

\item{l_MetO}{A list containing average external temperature data (assumed to include regional data).}
}
\value{
A list containing various solar parameter data frames.
}
\description{
This function calculates various solar parameters based on provided data sets for a dwelling.
}
