% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{evaluate_custom_archetype}
\alias{evaluate_custom_archetype}
\title{Evaluates a custom archetype}
\usage{
evaluate_custom_archetype(uid, dqu, sim = TRUE)
}
\arguments{
\item{uid}{Character string representing the UUID of the archetype to evaluate.}

\item{dqu}{Data frame containing archetype data.}

\item{sim}{Logical indicating whether to perform the EnergyPlus simulation.
Defaults to TRUE.}
}
\value{
The ID of the evaluated model.
}
\description{
This function evaluates a custom archetype based on the provided UUID. It
performs various steps including data subsetting, input assignment, IDF generation,
EnergyPlus simulation (if specified), and result analysis.
}
