% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{expand_output_resolution}
\alias{expand_output_resolution}
\title{Expands Output Resolution}
\usage{
expand_output_resolution(resolution)
}
\arguments{
\item{resolution}{The input resolution string.}
}
\value{
The expanded resolution string.
}
\description{
Converts various input resolution strings to
their corresponding standard formats. If the input resolution
is not found, the original string is returned.
}
\examples{
\dontrun{
expand_output_resolution("hourly")
}
}
