% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{extract_schedule_information}
\alias{extract_schedule_information}
\title{Extract schedule information from an IDF file}
\usage{
extract_schedule_information(idf)
}
\arguments{
\item{idf:}{A character vector representing the lines of an IDF file.}
}
\value{
: A character vector containing unique and sorted schedule names.
}
\description{
This function extracts schedule names from an IDF file.
}
