% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{extract_zone_temperature}
\alias{extract_zone_temperature}
\title{Extracts zone and environment temperatures from data}
\usage{
extract_zone_temperature(d)
}
\arguments{
\item{d}{A data frame containing temperature data.}
}
\value{
A data frame with extracted temperature information.
}
\description{
This function extracts zone operative temperature, mean air temperature,
radiant temperature, and outdoor and ground temperatures from a data frame.
}
