% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_biomass_store_objects}
\alias{generate_biomass_store_objects}
\title{Generates biomass storage objects.}
\usage{
generate_biomass_store_objects(params)
}
\arguments{
\item{params}{A list containing logical flags for various parameters.}
}
\value{
A list of IDF blocks representing biomass storage components.
}
\description{
This function creates IDF blocks for biomass storage components based on provided parameters.
}
