% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_dwelling_geometry}
\alias{generate_dwelling_geometry}
\title{Generate IDF list of thermal zones}
\usage{
generate_dwelling_geometry(l_d_info, l_layouts = t_layouts)
}
\arguments{
\item{l_d_info}{list of custom input parameters}

\item{l_layouts}{tupple of available layouts}

\item{l_def}{list of default cuboid inputs}
}
\value{
list of house geometry object as summary-card and idf
}
\description{
Generate IDF list of thermal zones
}
