% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{generate_energy_system}
\alias{generate_energy_system}
\title{Generate an idf block containing all energy system components}
\usage{
generate_energy_system(i_code, l_loop, d_zone)
}
\arguments{
\item{i_code:}{An integer representing the energy system code.}

\item{l_loop:}{A list containing loop parameters.}

\item{d_zone:}{A data frame containing zone data.}
}
\value{
: A list containing an idf of the energy system model, and
a summary of parameters
}
\description{
This function generates an energy system model based on
input codes, loop data, and zone data.
}
