% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{generate_heating_branch_list}
\alias{generate_heating_branch_list}
\title{Generate heating branch list}
\usage{
generate_heating_branch_list(k, d_zon, l_sys)
}
\arguments{
\item{k:}{Character representing the system type (e.g., 'wls', 'wld').}

\item{d_zon:}{Data frame containing zone information}

\item{l_sys:}{List of loop parameters}

\item{ig}{The input group.}
}
\value{
A data frame with columns:
\code{group} (system type)
\code{branch} (heating branch name)
\code{type} (heating branch function)
\code{loc} (heating branch location)
\code{link_in} (connection in pipes)
\code{link_out} (connection out pipes)
\code{has_pipe} (presence of adiabatic pipe)
\code{has_pump} (presence of pump device)

The assigned loop ID.
}
\description{
This function takes a system type (\code{k}), zone data
frame (\code{d_zon}), and loop parameters (\code{l_sys}), and generates a
data frame listing the heating branches for that system.

This function assigns a loop ID based on the provided input group.
}
\examples{
\dontrun{
k = define_heating_loops(l_loop$heating_codes$msh, l_in_loops)[4]
d_zon = d_zones
l_sys = l_in_loops
generate_heating_branch_list(k, d_zon, l_sys)
}
}
