% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus-wrapper.R
\name{generate_idf_multiple}
\alias{generate_idf_multiple}
\title{Generate multiple IDF files}
\usage{
generate_idf_multiple(s, profile_type = FALSE, res = "hourly")
}
\arguments{
\item{s}{A vector of simulation IDs.}

\item{profile_type}{Logical indicating whether to use fixed profiles (default: FALSE).}

\item{res}{The resolution of the IDF file (default: "hourly").}
}
\value{
None. The function generates IDF files and schedules as needed.
}
\description{
This function generates multiple IDF files based on the
provided simulation IDs. It checks for existing files and only generates
the missing ones.
}
\seealso{
Other EnergyPlus evaluation: 
\code{\link{convert_idf_to_json}()},
\code{\link{extract_simulation_outputs}()},
\code{\link{generate_idf}()},
\code{\link{read_idf_outcomes}()},
\code{\link{read_input_file}()},
\code{\link{read_input_summary}()},
\code{\link{run_energy_plus}()},
\code{\link{simulate_available_idfs}()}
}
\concept{EnergyPlus evaluation}
