% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_idf_zone_list}
\alias{generate_idf_zone_list}
\title{Generate IDF list of thermal zones}
\usage{
generate_idf_zone_list(l_zns)
}
\arguments{
\item{l_zns}{list of available zones}
}
\value{
list of thermal zones in idf format
}
\description{
Generate IDF list of thermal zones
}
