% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_list_of_surfaces}
\alias{generate_list_of_surfaces}
\title{Generate IDF surfaces}
\usage{
generate_list_of_surfaces(b_it, l_it, m_dm, l_dm, l_in, l_cr)
}
\arguments{
\item{b_it}{name of storey to be processed}

\item{l_it}{list of active house layouts}

\item{m_dm}{matrix of dimensions for each room}

\item{l_dm}{list of dimensions for each room}

\item{l_in}{list of custom input parameters}

\item{l_cr}{logical indication of core unit}
}
\description{
Generate IDF surfaces
}
\examples{
\dontrun{
generate_list_of_surfaces(b_sto[1], l_mir$left, m_dim, l_drm, b_cub, TRUE)
}
}
