% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_lzc_hosting_pv_surface}
\alias{generate_lzc_hosting_pv_surface}
\title{Create IDF block for Shading:Building based on surface data}
\usage{
generate_lzc_hosting_pv_surface(l_iter, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing loop properties (e.g., name, temperature).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a detailed
shading building using surface data (\code{l_iter}). It also uses supplementary
data (\code{l_sup}) for comments.
}
