% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_lzc_pv_generator}
\alias{generate_lzc_pv_generator}
\title{Create IDF block for Generator:Photovoltaic}
\usage{
generate_lzc_pv_generator(l_iter, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing loop properties (e.g., name, temperature).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a
photovoltaic generator using surface data (\code{l_iter}). It also uses
supplementary data (\code{l_sup}) for comments.
}
