% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_lzc_pv_load_centre_generators}
\alias{generate_lzc_pv_load_centre_generators}
\title{Create IDF block for ElectricLoadCenter:Generators}
\usage{
generate_lzc_pv_load_centre_generators(l_iter, i_pwr, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing loop properties (e.g., name, temperature).}

\item{i_pwr:}{A value indicating power capacity}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a
load centre generator using surface data (\code{l_iter}) and estimated power
capacity (\code{i_pwr}). It also uses supplementary data (\code{l_sup}) for comments.
}
