% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_lzc_wind_load_centre_distribution}
\alias{generate_lzc_wind_load_centre_distribution}
\title{Create IDF block for ElectricLoadCenter:Distribution}
\usage{
generate_lzc_wind_load_centre_distribution(i_battery, l_sup = l_idf_blocks)
}
\arguments{
\item{i_battery:}{A value indicating the presence of a battery}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a
load centre distribution object considering the presence of a battery
(\code{i_battery}). It also uses supplementary data (\code{l_sup}) for comments.
}
