% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_lzc_wind_turbine}
\alias{generate_lzc_wind_turbine}
\title{Create IDF block for Generator:WindTurbine}
\usage{
generate_lzc_wind_turbine(l_sup = l_idf_blocks)
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a
wind turbine generator object, It also uses supplementary
data (\code{l_sup}) for comments.
}
