% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{generate_performance_curves}
\alias{generate_performance_curves}
\title{Generates performance curves in IDF format}
\usage{
generate_performance_curves(
  l_used,
  l_curves = l_htg_curves,
  l_sup = l_idf_blocks
)
}
\arguments{
\item{l_used:}{A vector of curve names to be included.}

\item{l_curves:}{A list of curve data, indexed by curve type.}

\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector containing the generated IDF blocks.
}
\description{
Creates IDF blocks for performance curves based
on provided data.
}
