% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_pv_solar_objects}
\alias{generate_pv_solar_objects}
\title{Generates PV solar objects for a given set of houses.}
\usage{
generate_pv_solar_objects(
  houseIds,
  buildingGeometry,
  params,
  weatherData,
  includeBattery
)
}
\arguments{
\item{houseIds}{A vector of house IDs.}

\item{buildingGeometry}{A data frame containing building geometry information.}

\item{params}{A list containing logical flags for various parameters.}

\item{weatherData}{A data frame containing weather data.}

\item{includeBattery}{A logical flag indicating whether to include a battery.}
}
\value{
A list of IDF blocks representing PV components.
}
\description{
This function creates IDF blocks for PV components based on provided parameters.
}
