% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_room_layout}
\alias{generate_room_layout}
\title{Generate loom layout dictionary}
\usage{
generate_room_layout(idr, l_lay, l_axial, l_template = t_layouts)
}
\arguments{
\item{l_lay}{building layout info}

\item{l_axial}{list of axial dimensions}

\item{l_template}{tupple of templates of layouts}

\item{id}{label for selected story/floor}
}
\value{
a dictionary of layouts
}
\description{
Generate loom layout dictionary
}
