% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{generate_schedule__typical}
\alias{generate_schedule__typical}
\title{Generate Schedule with Typical Profile Variation}
\usage{
generate_schedule__typical(sel, var_int = 1, days = 365)
}
\arguments{
\item{sel}{A vector representing the typical profile.}

\item{var_int}{Numeric value controlling the amount of variation.
Higher values lead to larger deviations. (default: 1)}

\item{days}{Integer specifying the number of days in the schedule.
(default: 365)}
}
\value{
A numeric vector representing the generated schedule.
}
\description{
This function creates a schedule with a typical profile, incorporating
variation based on provided parameters.
}
