% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{generate_schedule_idf}
\alias{generate_schedule_idf}
\title{Generates IDF Schedule}
\usage{
generate_schedule_idf(ids, df_sched, df_apps)
}
\arguments{
\item{df_sched}{Data frame containing schedule data.}

\item{df_apps}{Data frame containing appliance data.}

\item{path_idf}{Character string specifying the output path for IDF files.}
}
\value{
A list containing the generated IDF schedule.
}
\description{
Creates an IDF schedule based on provided schedules and appliances data.
}
