% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_surfaces_door}
\alias{generate_surfaces_door}
\title{Generate IDF door objects for each of the surfaces}
\usage{
generate_surfaces_door(b_cub, b_glz, l_gl, l_dm)
}
\arguments{
\item{b_cub}{information card of house properties}

\item{b_glz}{list of active rooms}

\item{l_gl}{list of glazing dimensions}

\item{l_dm}{list of home info}
}
\value{
list of door block by unit/storey/dict,idf
}
\description{
Generate IDF door objects for each of the surfaces
}
\examples{
\dontrun{
generate_surfaces_door(l_cuboid, l_glazing, l_glazed, l_dim)
}




}
