% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_surfaces_glazing}
\alias{generate_surfaces_glazing}
\title{Generate IDF fenestration objects for each of the surfaces}
\usage{
generate_surfaces_glazing(b_cub, b_glz, l_gl, l_lg, l_bl)
}
\arguments{
\item{b_cub}{information card of house properties}

\item{b_glz}{list of active rooms}

\item{l_gl}{list of glazing dimensions}

\item{l_lg}{list of home info}

\item{l_bl}{list of layout templates}
}
\value{
list of fenestration blocks by unit/storey/dict,idf
}
\description{
Generate IDF fenestration objects for each of the surfaces
}
\examples{
\dontrun{
generate_surfaces_glazing(l_cuboid, l_glazing, l_glazed, l_glazde, l_blocks)
}
}
