% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{generate_surfaces_walls}
\alias{generate_surfaces_walls}
\title{Generate IDF objects for each of the surfaces}
\usage{
generate_surfaces_walls(b_cub, l_rly, l_drm, l_tpl = t_layouts)
}
\arguments{
\item{b_cub}{information card of house properties}

\item{l_rly}{list of active rooms}

\item{l_drm}{list of dimensions}

\item{l_tpl}{list of layout templates}
}
\value{
list of opaque blocks by unit/storey/dict,idf
}
\description{
Generate IDF objects for each of the surfaces
}
\examples{
\dontrun{
 generate_surfaces_walls(l_cuboid, l_room_layouts, l_dim)
}
}
