% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{generate_timeslot_table}
\alias{generate_timeslot_table}
\title{Generate Timeslot Table}
\usage{
generate_timeslot_table(
  year = the$GLOBAL_PAR$simulation_year,
  resolution = "10 min"
)
}
\arguments{
\item{year}{The year for which to generate time slots (default:
the$GLOBAL_PAR$simulation_year)}

\item{resolution}{The time slot interval (default: "10 min").}
}
\value{
A data.frame with the following columns:
\itemize{
\item tid: Unique identifier for the time slot.
\item tid.from: Starting time of the time slot (HH:MM format).
\item tid.to: Ending time of the time slot (HH:MM format).
\item tid.slot: Combined string representation of starting and ending times (HH:MM-HH:MM).
}
}
\description{
Creates a data.frame defining time slots for a given year.
}
