% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{generate_water_service}
\alias{generate_water_service}
\title{Generates water service IDF blocks.}
\usage{
generate_water_service(l_zone_use, d_branch, n_occ)
}
\arguments{
\item{l_zone_use}{A list containing zone usage information.}

\item{n_occ}{Number of occupants.}
}
\value{
A list containing the IDF blocks and water service summary data.
}
\description{
This function creates IDF blocks for water services based on zone usage and number of occupants.
}
