% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_water_solar_objects}
\alias{generate_water_solar_objects}
\title{Generates solar water heating objects.}
\usage{
generate_water_solar_objects(params)
}
\arguments{
\item{params}{A list containing logical flags for various parameters.}
}
\value{
A list of IDF blocks representing solar water heating components.
}
\description{
This function creates IDF blocks for solar water heating components based on provided parameters.
}
