% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{generate_wind_turbine_objects}
\alias{generate_wind_turbine_objects}
\title{Generates wind turbine objects for a given set of houses.}
\usage{
generate_wind_turbine_objects(houseIds, params, weatherData, includeBattery)
}
\arguments{
\item{houseIds}{A vector of house IDs.}

\item{params}{A list containing logical flags for various parameters.}

\item{weatherData}{A data frame containing weather data.}

\item{includeBattery}{A logical flag indicating whether to include a battery.}
}
\value{
A list of IDF blocks representing wind turbine components.
}
\description{
This function creates IDF blocks for wind turbine components based on provided parameters.
}
