% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{get_aacode}
\alias{get_aacode}
\title{Retrieve IDF Codes}
\usage{
get_aacode(codes)
}
\arguments{
\item{codes}{The input codes for retrieving IDF codes.}
}
\value{
A vector of IDF codes.
}
\description{
Extracts IDF codes from a reference data frame based on the
specified mode and input codes.
}
\examples{
get_aacode(2)
}
