% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_effective_zones}
\alias{get_effective_zones}
\title{Retrieve Effective Zones for Housing Code}
\usage{
get_effective_zones(housing_code, data_list)
}
\arguments{
\item{housing_code}{The housing code.}

\item{data_list}{A list containing data for processing.}
}
\value{
A list with two elements:
\itemize{
\item \code{info}: A data frame containing archetype parameters.
\item \code{zones}: A list of effective zones.
}
}
\description{
This function calculates effective zones based on the provided housing code and data.
It retrieves necessary information from the \code{data_list}, including complementary data,
archetype parameters, and zonal definitions.
}
