% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_enhub_location}
\alias{get_enhub_location}
\title{Gets EnHub Project Location}
\usage{
get_enhub_location()
}
\value{
The normalized path to the EnHub project directory.
}
\description{
Determines the current EnHub project location based
on the operating system.
}
\details{
This function retrieves the current working directory (project location)
using platform-specific methods. It normalizes the path for consistency
across operating systems.
}
\examples{
\dontrun{
get_enhub_location() # Retrieves the EnHub directory path
}
}
\seealso{
\code{\link[=get_ep_location]{get_ep_location()}} for a function that checks the location
of the EnergyPlus (available) folder.
}
