% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_ep_location}
\alias{get_ep_location}
\title{Gets EnergyPlus Binary Location}
\usage{
get_ep_location()
}
\value{
The path to the EnergyPlus binary directory.
}
\description{
Determines the location of the EnergyPlus binary file.
}
\details{
This function retrieves the EnergyPlus binary directory path from
the \code{the$.enhub.config} settings. It checks if the directory exists and
prints a message accordingly.
}
\note{
The function currently only supports retrieving the path from the
configuration and does not implement platform-specific search logic.
\itemize{
\item Windows: \verb{cmd >> dir /b /s .\\\\*RunEPlus.bat}
\item macOS: \verb{mdfind -name runenergyplus}
\item Linux: \verb{find / -name runenergyplus}
}
}
\examples{
\dontrun{
get_ep_location() # Retrieves the EnergyPlus binary directory path
}
}
\seealso{
\code{\link[=get_enhub_location]{get_enhub_location()}} for a function that checks the location
of the project directory.
}
