% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typology.R
\name{get_input_set}
\alias{get_input_set}
\title{Creates a List of Variables for Subsetting EHS Data}
\usage{
get_input_set(data_frame = combine_ehs())
}
\arguments{
\item{data_frame}{A data frame containing EHS data (defaults to the output of \code{combine_ehs()}).}
}
\value{
A list containing information on potential subsetting variables.
}
\description{
This function identifies potential variables for subsetting EHS data based on their types (categorical, numeric, logical).
It retrieves available levels for categorical variables, lists possible values for reference variables,
and provides summary statistics for numeric variables.
}
