% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_jsons}
\alias{get_jsons}
\title{Loads Multiple JSON Files}
\usage{
get_jsons(path_tbl)
}
\arguments{
\item{path_tbl}{The path to the directory containing the JSON files.}
}
\value{
A list of JSON objects, where each object corresponds to a JSON file.
}
\description{
Loads multiple JSON files from a specified directory path.
}
\details{
This function loads multiple JSON files from a specified directory path.
For each JSON file, it reads the data using \code{fromJSON}. The resulting
JSON objects are stored in a list with names derived from the file names.
}
