% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_tables}
\alias{get_tables}
\title{Loads Multiple CSV Tables}
\usage{
get_tables(path_tbl)
}
\arguments{
\item{path_tbl}{The path to the directory containing the CSV files.}

\item{l_f}{The file path of the CSV file.}
}
\value{
A list of data frames, where each data frame corresponds to a CSV file.

A tibble containing the loaded data.
}
\description{
Loads multiple CSV tables from a specified directory path.
}
\details{
This function loads multiple CSV files from a specified directory path.
For each CSV file, it reads the data, converts column names to
lowercase and replaces dots with underscores, and converts the data
frame to a tibble. The resulting data frames are stored in a list with
names derived from the file names.
}
