% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{idf_built_status}
\alias{idf_built_status}
\title{IDF Built Status}
\usage{
idf_built_status(label, extra_info = NULL)
}
\arguments{
\item{label}{A character string representing the label
for the status message.}

\item{extra_info}{An optional character string for
additional information to be appended to the status message.}
}
\value{
The function returns the formatted status message.
}
\description{
This function generates a formatted message
indicating the status of an IDF project configuration.
}
\examples{
\dontrun{
idf_built_status("IDF project configuration")
idf_built_status("IDF project configuration", "Additional info")
}
}
