% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialiser.R
\name{init_project}
\alias{init_project}
\title{Initialize the Project}
\usage{
init_project(core_data = "ehs", core_mode = "def", fresh_setup = TRUE)
}
\arguments{
\item{core_data}{A character string indicating the type of data.}

\item{core_mode}{A character string indicating the mode of operation.}

\item{fresh_setup}{A logical value indicating whether to perform a fresh setup.}
}
\value{
Invisible NULL.
}
\description{
This function sets up the \code{R} environment by loading necessary
libraries, setting up the session, and loading or generating data as needed.
}
