% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-format.R
\name{iterate_on_list}
\alias{iterate_on_list}
\title{Iterate Over List of Parameters}
\usage{
iterate_on_list(lpar, com)
}
\arguments{
\item{lpar}{A list of parameters.}

\item{com}{A vector of comments.}
}
\value{
A character vector of IDF comments.
}
\description{
This function iterates over a list of parameters and creates IDF comments.
}
\examples{
\dontrun{
iterate_on_list(list("param1", "param2"), c("comment1", "comment2"))
}
}
\seealso{
Other idf-formatting: 
\code{\link{add_idf_comment}()},
\code{\link{add_idf_comment_custom}()},
\code{\link{determine_end_of_idf_block}()},
\code{\link{iterate_on_output_list}()},
\code{\link{make_idf_block}},
\code{\link{make_idf_block_header}()},
\code{\link{make_idf_block_title}()}
}
\concept{idf-formatting}
