% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_activities}
\alias{l_activities}
\title{Activity Data List}
\format{
A list of 3 data frames:
\describe{
\item{types}{A data frame with 290 observations of 6 variables:
\describe{
\item{id}{Integer: Activity ID}
\item{name}{Character: Activity name}
\item{in2000}{Logical: Indicator if the activity was present in the year 2000}
\item{in2015}{Logical: Indicator if the activity was present in the year 2015}
\item{inboth}{Logical: Indicator if the activity was present in both years}
\item{group}{Character: Activity group}
}
}
\item{intensity_abs}{A data frame with 25 observations of 3 variables:
\describe{
\item{hgh}{Numeric: High intensity measure}
\item{nom}{Integer: Nominal intensity measure}
\item{low}{Numeric: Low intensity measure}
}
}
\item{intensity_rel}{A data frame with 24 observations of 3 variables:
\describe{
\item{hgh}{Integer: High relative intensity measure}
\item{nom}{Numeric: Nominal relative intensity measure}
\item{low}{Numeric: Low relative intensity measure}
}
}
}
}
\usage{
l_activities
}
\description{
A list containing three data frames related to activities, their types, and intensity measures.
}
\examples{
\dontrun{
# Access the list
l_activities

# Access the 'types' data frame
l_activities$types

# Access the 'intensity_abs' data frame
l_activities$intensity_abs

# Access the 'intensity_rel' data frame
l_activities$intensity_rel
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
