% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_activity_groups}
\alias{l_activity_groups}
\title{Activity Groups Data}
\format{
A list of one data frame with 10 observations on the following 2 variables:
\describe{
\item{id}{Integer. The unique identifier for each activity group.}
\item{name}{Character. The name of the activity group.}
}
}
\usage{
l_activity_groups
}
\description{
This dataset contains information about different activity groups.
Each activity group is represented by an ID and a name.
}
\examples{
\dontrun{
# Displaying the whole list
l_activity_groups
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
