% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_appliance_rating}
\alias{l_appliance_rating}
\title{Appliance Rating Data}
\format{
A list of one data frame with 11 observations on the following 4 variables:
\describe{
\item{colour_carbon}{Character. The color code representing the carbon rating of the appliance.}
\item{colour_energy}{Character. The color code representing the energy rating of the appliance.}
\item{id}{Integer. The unique identifier for each appliance rating.}
\item{rating}{Character. The rating of the appliance (e.g., "A++", "A+", "A", etc.).}
}
}
\usage{
l_appliance_rating
}
\description{
This dataset contains information about appliance ratings, including their carbon and energy colours, IDs, and ratings.
}
\examples{
\dontrun{
# Access the first appliance's carbon color
l_appliance_rating$colour_carbon[1]

# Access the energy color of the second appliance
l_appliance_rating$colour_energy[2]

# Access the ID of the third appliance
l_appliance_rating$id[3]

# Access the rating of the fourth appliance
l_appliance_rating$rating[4]
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
