% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--banners.R
\docType{data}
\name{l_banners}
\alias{l_banners}
\title{List of banners}
\format{
A list of text-based banners used for improved display of the output.
\itemize{
\item \code{main}: Main banner text
\item \code{head}: Header banner text
\item \code{mans}: Mans banner text
\item \code{para}: Parametrics banner text
\item \code{maps}: UK map banner text
\item \code{wflw}: Workflow banner text
}
}
\usage{
l_banners
}
\description{
This list contains various banner texts read from different files.
}
\examples{
\dontrun{
# Displaying the whole list
l_banners
}
}
\keyword{datasets}
