% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_comfort_states}
\alias{l_comfort_states}
\title{Dictionary with comfort states information}
\format{
A list containing datasets related to thermal sensation categories.
\itemize{
\item fanger: A character vector representing Fanger's thermal sensation scale.
\itemize{
\item cold: Feeling cold.
\item cool: Feeling cool.
\item slightly cool: Feeling slightly cool.
\item neutral: Feeling neutral.
\item slightly warm: Feeling slightly warm.
\item warm: Feeling warm.
\item hot: Feeling hot.
}
\item general: A character vector representing general thermal sensation categories.
\itemize{
\item a bit too cold: Feeling a bit too cold.
\item ok: Feeling comfortable.
\item a bit too hot: Feeling a bit too hot.
}
}
}
\usage{
l_comfort_states
}
\description{
This object contains a dictionary with comfort states information.
}
\examples{
\dontrun{
# Displaying the whole list
l_comfort_states
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
