% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_epw_header_units}
\alias{l_epw_header_units}
\title{EPW Header Units Data}
\format{
A list of one data frame with 35 observations on the following 3 variables:
\describe{
\item{description}{Character. The description of the header unit (e.g., "Year", "Month", "Day").}
\item{name}{Character. The name of the header unit (e.g., "Year", "Month", "Day").}
\item{units}{Character. The unit of measurement for the header unit (e.g., "year", "month", "day").}
}
}
\usage{
l_epw_header_units
}
\description{
This dataset contains the header units for EPW (EnergyPlus Weather) files.
Each header unit includes a description, name, and unit of measurement.
}
\examples{
\dontrun{
# Access the description of the first header unit
l_epw_header_units$description[1]

# Access the name of the second header unit
l_epw_header_units$name[2]

# Access the unit of the third header unit
l_epw_header_units$units[3]

# Access the description, name, and unit of the fourth header unit
l_epw_header_units$description[4]
l_epw_header_units$name[4]
l_epw_header_units$units[4]
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
