% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--hlp.R
\docType{data}
\name{l_HLP}
\alias{l_HLP}
\title{Dictionary with Heat Loss Parameters (HLP) information}
\format{
List of tibbles containing construction and material data for building energy modelling.
\itemize{
\item HLP_constructionsOpaque: Tibble containing data on opaque constructions. This tibble has 38 rows and 8 columns. Each row represents a different type of opaque construction.
\itemize{
\item name: The name of the construction (e.g., "stone: granite or whin (as built)").
\item outside_layer: The material used for the outside layer (e.g., "200 mm granite block").
\item layer_2: The material used for the second layer (e.g., "13 mm dense plaster").
\item layer_3: The material used for the third layer (e.g., "-").
\item layer_4: The material used for the fourth layer (e.g., "-").
\item layer_5: The material used for the fifth layer (e.g., "-").
\item code: A unique code for the construction (e.g., 10001).
\item source: The source of the data (e.g., "CIBSE Guide A").
}
\item HLP_constructionsTranslucent: Tibble containing data on translucent constructions. This tibble has 219 rows and 10 columns. Each row represents a different type of translucent construction.
\itemize{
\item name: The name of the construction (e.g., "Composite 2-Core Filled Concrete Block Insulated").
\item outside_layer: The material used for the outside layer (e.g., "Composite 2-Core Filled Concrete Block Insulated #3").
\item layer_2: The material used for the second layer (e.g., "Composite 2-Core Filled Concrete Block Insulated #2").
\item layer_3: The material used for the third layer (e.g., "Composite 2-Core Filled Concrete Block Insulated #1").
\item layer_4: The material used for the fourth layer (e.g., "-").
\item layer_5: The material used for the fifth layer (e.g., "-").
\item layer_6: The material used for the sixth layer (e.g., "-").
\item layer_7: The material used for the seventh layer (e.g., "-").
\item code: A unique code for the construction (e.g., 20001).
\item source: The source of the data (e.g., "EnergyPlus Datasets").
}
\item HLP_materialsOpaque: Tibble containing data on opaque materials. This tibble has 298 rows and 13 columns. Each row represents a different type of opaque material.
\itemize{
\item type: The type of material (e.g., "Material:AirGap").
\item name: The name of the material (e.g., "F04 Wall air space resistance").
\item roughness: The roughness of the material (e.g., "-").
\item thickness__m_: The thickness of the material in meters (e.g., "-").
\item conductivity__w_m_k_: The thermal conductivity of the material in W/m·K (e.g., "-").
\item density__kg_m3_: The density of the material in kg/m³ (e.g., "-").
\item specific_heat__j_kg_k_: The specific heat of the material in J/kg·K (e.g., "-").
\item thermal_resistance__m2_k_w_: The thermal resistance of the material in m²·K/W (e.g., "0.15").
\item thermal_absorptance: The thermal absorptance of the material (e.g., "-").
\item solar_absorptance: The solar absorptance of the material (e.g., "-").
\item visible_absorptance: The visible absorptance of the material (e.g., "-").
\item code: A unique code for the material (e.g., 1).
\item source: The source of the data (e.g., "EnergyPlus Datasets").
}
\item HLP_materialsTranslucent: Tibble containing data on translucent materials. This tibble has 128 rows and 72 columns. Each row represents a different type of translucent material.
\itemize{
\item type: The type of material (e.g., "Material").
\item name: The name of the material (e.g., "Composite 2-Core Filled Concrete Block Insulated #1").
\item airflow_permeability__dimensionless_: The airflow permeability of the material (e.g., "-").
\item angle_of_resolution_for_screen_transmittance_output_map__deg_: The angle of resolution for screen transmittance output map in degrees (e.g., "-").
\item back_side_infrared_hemispherical_emissivity: The back side infrared hemispherical emissivity of the material (e.g., "-").
\item back_side_slat_beam_solar_reflectance: The back side slat beam solar reflectance of the material (e.g., "-").
\item back_side_slat_beam_visible_reflectance: The back side slat beam visible reflectance of the material (e.g., "-").
\item back_side_slat_diffuse_solar_reflectance__dimensionless_: The back side slat diffuse solar reflectance of the material (e.g., "-").
\item back_side_slat_diffuse_visible_reflectance__dimensionless_: The back side slat diffuse visible reflectance of the material (e.g., "-").
\item back_side_slat_infrared_hemispherical_emissivity: The back side slat infrared hemispherical emissivity of the material (e.g., "-").
\item back_side_solar_reflectance_at_normal_incidence: The back side solar reflectance at normal incidence of the material (e.g., "-").
\item back_side_visible_reflectance_at_normal_incidence: The back side visible reflectance at normal incidence of the material (e.g., "-").
\item blind_bottom_opening_multiplier: The blind bottom opening multiplier of the material (e.g., "-").
\item blind_left_side_opening_multiplier: The blind left side opening multiplier of the material (e.g., "-").
\item blind_right_side_opening_multiplier: The blind right side opening multiplier of the material (e.g., "-").
\item blind_to_glass_distance__m_: The blind to glass distance in meters (e.g., "-").
\item blind_top_opening_multiplier: The blind top opening multiplier of the material (e.g., "-").
\item bottom_opening_multiplier__dimensionless_: The bottom opening multiplier of the material (e.g., "-").
\item conductivity__w_m_k_: The thermal conductivity of the material in W/m·K (e.g., "0.88").
\item density__kg_m3_: The density of the material in kg/m³ (e.g., "1774.75").
\item diffuse_solar_reflectance: The diffuse solar reflectance of the material (e.g., "-").
\item diffuse_visible_reflectance: The diffuse visible reflectance of the material (e.g., "-").
\item front_side_infrared_hemispherical_emissivity: The front side infrared hemispherical emissivity of the material (e.g., "-").
\item front_side_slat_beam_solar_reflectance: The front side slat beam solar reflectance of the material (e.g., "-").
\item front_side_slat_beam_visible_reflectance: The front side slat beam visible reflectance of the material (e.g., "-").
\item front_side_slat_diffuse_solar_reflectance: The front side slat diffuse solar reflectance of the material (e.g., "-").
\item front_side_slat_diffuse_visible_reflectance: The front side slat diffuse visible reflectance of the material (e.g., "-").
\item front_side_slat_infrared_hemispherical_emissivity: The front side slat infrared hemispherical emissivity of the material (e.g., "-").
\item front_side_solar_reflectance_at_normal_incidence: The front side solar reflectance at normal incidence of the material (e.g., "-").
\item front_side_visible_reflectance_at_normal_incidence: The front side visible reflectance at normal incidence of the material (e.g., "-").
\item gas_type: The type of gas used in the material (e.g., "-").
\item infrared_hemispherical_emissivity__dimensionless_: The infrared hemispherical emissivity of the material (e.g., "-").
\item infrared_transmittance__dimensionless_: The infrared transmittance of the material (e.g., "-").
\item infrared_transmittance_at_normal_incidence: The infrared transmittance at normal incidence of the material (e.g., "-").
\item left_side_opening_multiplier__dimensionless_: The left side opening multiplier of the material (e.g., "-").
\item optical_data_type: The optical data type of the material (e.g., "-").
\item reflected_beam_transmittance_accounting_method: The reflected beam transmittance accounting method of the material (e.g., "-").
\item right_side_opening_multiplier__dimensionless_: The right side opening multiplier of the material (e.g., "-").
\item roughness: The roughness of the material (e.g., "Mediumrough").
\item screen_material_diameter__m_: The screen material diameter in meters (e.g., "-").
\item screen_material_spacing__m_: The screen material spacing in meters (e.g., "-").
\item screen_to_glass_distance__m_: The screen to glass distance in meters (e.g., "-").
\item shade_to_glass_distance: The shade to glass distance of the material (e.g., "-").
\item slat_angle__deg_: The slat angle in degrees (e.g., "-").
\item slat_beam_solar_transmittance: The slat beam solar transmittance of the material (e.g., "-").
\item slat_beam_visible_transmittance: The slat beam visible transmittance of the material (e.g., "-").
\item slat_conductivity__w_m_k_: The slat conductivity of the material in W/m·K (e.g., "-").
\item slat_diffuse_solar_transmittance: The slat diffuse solar transmittance of the material (e.g., "-").
\item slat_diffuse_visible_transmittance: The slat diffuse visible transmittance of the material (e.g., "-").
\item slat_infrared_hemispherical_transmittance: The slat infrared hemispherical transmittance of the material (e.g., "-").
\item slat_orientation: The slat orientation of the material (e.g., "-").
\item slat_separation__m_: The slat separation in meters (e.g., "-").
\item slat_thickness__m_: The slat thickness in meters (e.g., "-").
\item slat_width__m_: The slat width in meters (e.g., "-").
\item solar_absorptance: The solar absorptance of the material (e.g., "0.7").
\item solar_reflectance__dimensionless_: The solar reflectance of the material (e.g., "-").
\item solar_transmittance__dimensionless_: The solar transmittance of the material (e.g., "-").
\item solar_transmittance_at_normal_incidence: The solar transmittance at normal incidence of the material (e.g., "-").
\item specific_heat: The specific heat of the material (e.g., "838").
\item thermal_absorptance: The thermal absorptance of the material (e.g., "0.9").
\item thermal_hemispherical_emissivity__dimensionless_: The thermal hemispherical emissivity of the material (e.g., "-").
\item thickness__m_: The thickness of the material in meters (e.g., "0.079").
\item top_opening_multiplier__dimensionless_: The top opening multiplier of the material (e.g., "-").
\item visible_absorptance: The visible absorptance of the material (e.g., "0.7").
\item visible_reflectance__dimensionless_: The visible reflectance of the material (e.g., "-").
\item visible_transmittance__dimensionless_: The visible transmittance of the material (e.g., "-").
\item visible_transmittance_at_normal_incidence: The visible transmittance at normal incidence of the material (e.g., "-").
\item window_glass_spectral_data_set_name: The window glass spectral data set name of the material (e.g., "-").
\item minimum_slat_angle__deg_: The minimum slat angle in degrees (e.g., "-").
\item maximum_slat_angle__deg_: The maximum slat angle in degrees (e.g., "-").
\item code: A unique code for the material (e.g., 1001).
\item source: The source of the data (e.g., "EnergyPlus Datasets").
}
}
}
\usage{
l_HLP
}
\description{
This object contains a dictionary with Heat Loss Parameters (HLP) information
}
\examples{
\dontrun{
# Displaying the whole list
l_HLP
}
}
\keyword{datasets}
