% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_housing_groups}
\alias{l_housing_groups}
\title{Housing Groups Data}
\format{
A list of 15 elements, each containing a data frame with the following variables:
\describe{
\item{activity}{A data frame with 12 observations on 3 variables: \code{id} (character), \code{name} (character), \code{AC} (integer).}
\item{category}{A data frame with 12 observations on 3 variables: \code{id} (character), \code{name} (character), \code{CA} (integer).}
\item{composition}{A data frame with 3 observations on 3 variables: \code{id} (character), \code{name} (character), \code{CH} (integer).}
\item{dwellingTenure}{A data frame with 2 observations on 3 variables: \code{id} (character), \code{name} (character), \code{TN} (integer).}
\item{dwellingType}{A data frame with 5 observations on 3 variables: \code{id} (character), \code{name} (character), \code{TY} (integer).}
\item{fuelPoverty}{A data frame with 3 observations on 3 variables: \code{id} (character), \code{name} (character), \code{FP} (integer).}
\item{householdComposition}{A data frame with 5 observations on 3 variables: \code{id} (character), \code{name} (character), \code{HH} (integer).}
\item{HRPAge}{A data frame with 3 observations on 3 variables: \code{id} (character), \code{name} (character), \code{AX} (integer).}
\item{HRPEmploymentStatus}{A data frame with 3 observations on 3 variables: \code{id} (character), \code{name} (character), \code{WS} (integer).}
\item{individualAge}{A data frame with 5 observations on 3 variables: \code{id} (character), \code{name} (character), \code{YR} (integer).}
\item{individualGender}{A data frame with 2 observations on 3 variables: \code{id} (character), \code{name} (character), \code{GR} (integer).}
\item{lights}{A data frame with 9 observations on 4 variables: \code{id} (character), \code{name} (character), \code{power} (numeric), \code{LT} (integer).}
\item{relationships}{A data frame with 24 observations on 2 variables: \code{id} (integer), \code{name} (character).}
\item{rooms}{A data frame with 33 observations on 3 variables: \code{id} (character), \code{name} (character), \code{ZN} (integer).}
\item{typology}{A data frame with 4 observations on 3 variables: \code{id} (character), \code{name} (character), \code{LY} (integer).}
}
}
\usage{
l_housing_groups
}
\description{
This dataset contains various housing groups, each represented by a data frame with specific attributes.
The groups include activities, categories, compositions, dwelling tenures, dwelling types, fuel poverty statuses, household compositions, HRP ages, HRP employment statuses, individual ages, individual genders, lights, relationships, rooms, and typologies.
}
\examples{
\dontrun{
# Access the first activity's name
l_housing_groups$activity$name[1]

# Access the category ID of the second category
l_housing_groups$category$id[2]

# Access the composition name of the third composition
l_housing_groups$composition$name[3]

# Access the dwelling tenure name of the first dwelling tenure
l_housing_groups$dwellingTenure$name[1]

# Access the power of the first light
l_housing_groups$lights$power[1]
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
