% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_htg_curves}
\alias{l_htg_curves}
\title{Dictionary with IDF modelling curves}
\format{
A list containing various curve datasets used for different types of performance and efficiency calculations.
\itemize{
\item Curve:Biquadratic: A list containing biquadratic curve datasets.
\itemize{
\item ASHP HighT CAPFT: A list containing coefficients and parameters for the ASHP High Temperature Capacity Factor.
\itemize{
\item coefficient1_constant: 0.975
\item coefficient2_x: 0.0204
\item coefficient3_x_2: 0.0003
\item coefficient4_y: -0.0049
\item coefficient5_y_2: 0
\item coefficient6_x_y: 0
\item input_unit_type_for_x: "Temperature"
\item input_unit_type_for_y: "Temperature"
\item maximum_value_of_x: 20
\item maximum_value_of_y: 80
\item minimum_value_of_x: -15
\item minimum_value_of_y: 35
\item output_unit_type: "Dimensionless"
}
\item ASHP HighT COPFT: A list containing coefficients and parameters for the ASHP High Temperature Coefficient of Performance Factor.
\itemize{
\item coefficient1_constant: 1.9
\item coefficient2_x: 0.0359
\item coefficient3_x_2: 0.0003
\item coefficient4_y: -0.0343
\item coefficient5_y_2: 0.0002
\item coefficient6_x_y: -0.0003
\item input_unit_type_for_x: "Temperature"
\item input_unit_type_for_y: "Temperature"
\item maximum_value_of_x: 20
\item maximum_value_of_y: 80
\item minimum_value_of_x: -15
\item minimum_value_of_y: 35
\item output_unit_type: "Dimensionless"
}
\item Boiler Efficiency: A list containing coefficients and parameters for boiler efficiency.
\itemize{
\item coefficient1_constant: 1
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item coefficient4_y: 0
\item coefficient5_y_2: 0
\item coefficient6_x_y: 0
\item maximum_value_of_x: 1
\item maximum_value_of_y: 1
\item minimum_value_of_x: 0
\item minimum_value_of_y: 0
}
\item CondensingBoilerEff: A list containing coefficients and parameters for condensing boiler efficiency.
\itemize{
\item coefficient1_constant: 1.12
\item coefficient2_x: 0.015
\item coefficient3_x_2: -0.026
\item coefficient4_y: 0
\item coefficient5_y_2: -1.4e-06
\item coefficient6_x_y: -0.0015
\item maximum_value_of_x: 1
\item maximum_value_of_y: 85
\item minimum_value_of_x: 0.1
\item minimum_value_of_y: 30
}
\item DefrostEIR: A list containing coefficients and parameters for defrost energy input ratio.
\itemize{
\item coefficient1_constant: 0.153
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item coefficient4_y: 0
\item coefficient5_y_2: 0
\item coefficient6_x_y: 0
\item maximum_value_of_x: 100
\item maximum_value_of_y: 100
\item minimum_value_of_x: -100
\item minimum_value_of_y: -100
}
}
\item Curve:Cubic: A list containing cubic curve datasets.
\itemize{
\item DefaultFanEffRatioCurve: A list containing coefficients and parameters for the default fan efficiency ratio curve.
\itemize{
\item coefficient1_constant: 0.339
\item coefficient2_x: 1.73
\item coefficient3_x_2: -1.49
\item coefficient4_x_3: 0.428
\item maximum_curve_output: 1
\item maximum_value_of_x: 1.5
\item minimum_curve_output: 0.3
\item minimum_value_of_x: 0.5
}
\item New style low temperature boiler: A list containing coefficients and parameters for the new style low temperature boiler.
\itemize{
\item coefficient1_constant: 0.839
\item coefficient2_x: 0.133
\item coefficient3_x_2: -0.17
\item coefficient4_x_3: 0.0475
\item maximum_value_of_x: 1
\item minimum_value_of_x: 0.1
}
}
\item Curve:Exponent: A list containing exponential curve datasets.
\itemize{
\item DefaultFanPowerRatioCurve: A list containing coefficients and parameters for the default fan power ratio curve.
\itemize{
\item coefficient1_constant: 0
\item coefficient2_constant: 1
\item coefficient3_constant: 3
\item maximum_curve_output: 1.5
\item maximum_value_of_x: 1.5
\item minimum_curve_output: 0.01
\item minimum_value_of_x: 0
}
}
\item Curve:QuadLinear: A list containing quadlinear curve datasets.
\itemize{
\item WWHPHeatCapCurve: A list containing coefficients and parameters for the water-to-water heat pump heat capacity curve.
\itemize{
\item coefficient1_constant: -1.6
\item coefficient2_w: -0.983
\item coefficient3_x: 3.25
\item coefficient4_y: 0.1
\item coefficient5_z: 0.0798
\item maximum_value_of_w: 100
\item maximum_value_of_x: 100
\item maximum_value_of_y: 100
\item maximum_value_of_z: 100
\item minimum_value_of_w: -100
\item minimum_value_of_x: -100
\item minimum_value_of_y: -100
\item minimum_value_of_z: -100
}
\item WWHPHeatPowCurve: A list containing coefficients and parameters for the water-to-water heat pump heat power curve.
\itemize{
\item coefficient1_constant: -2.68
\item coefficient2_w: 2.69
\item coefficient3_x: 0.393
\item coefficient4_y: 0.0109
\item coefficient5_z: 0.0149
\item maximum_value_of_w: 100
\item maximum_value_of_x: 100
\item maximum_value_of_y: 100
\item maximum_value_of_z: 100
\item minimum_value_of_w: -100
\item minimum_value_of_x: -100
\item minimum_value_of_y: -100
\item minimum_value_of_z: -100
}
}
\item Curve:Quadratic: A list containing quadratic curve datasets.
\itemize{
\item HPWHHeatingCOPFAirFlow: A list containing coefficients and parameters for the heat pump water heater heating COP as a function of air flow.
\itemize{
\item coefficient1_constant: 1
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item input_unit_type_for_x: "Dimensionless"
\item maximum_value_of_x: 2
\item minimum_value_of_x: 0
\item output_unit_type: "Dimensionless"
}
\item HPWHHeatingCOPFWaterFlow: A list containing coefficients and parameters for the heat pump water heater heating COP as a function of water flow.
\itemize{
\item coefficient1_constant: 1
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item input_unit_type_for_x: "Dimensionless"
\item maximum_value_of_x: 2
\item minimum_value_of_x: 0
\item output_unit_type: "Dimensionless"
}
\item HPWHHeatingCapFAirFlow: A list containing coefficients and parameters for the heat pump water heater heating capacity as a function of air flow.
\itemize{
\item coefficient1_constant: 1
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item input_unit_type_for_x: "Dimensionless"
\item maximum_value_of_x: 2
\item minimum_value_of_x: 0
\item output_unit_type: "Dimensionless"
}
\item HPWHHeatingCapFWaterFlow: A list containing coefficients and parameters for the heat pump water heater heating capacity as a function of water flow.
\itemize{
\item coefficient1_constant: 1
\item coefficient2_x: 0
\item coefficient3_x_2: 0
\item input_unit_type_for_x: "Dimensionless"
\item maximum_value_of_x: 2
\item minimum_value_of_x: 0
\item output_unit_type: "Dimensionless"
}
\item HPWHPLFFPLR: A list containing coefficients and parameters for the heat pump water heater part load fraction as a function of part load ratio.
\itemize{
\item coefficient1_constant: 0.75
\item coefficient2_x: 0.25
\item coefficient3_x_2: 0
\item input_unit_type_for_x: "Dimensionless"
\item maximum_value_of_x: 1
\item minimum_value_of_x: 0
\item output_unit_type: "Dimensionless"
}
}
}
}
\usage{
l_htg_curves
}
\description{
This object contains a dictionary with IDF modelling curves
}
\examples{
\dontrun{
# Displaying the whole list
l_htg_curves
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
