% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_idf_algorithm}
\alias{l_idf_algorithm}
\title{IDF Algorithm Data}
\format{
A list of one element containing a character vector with 4 observations:
\describe{
\item{algorithm}{Character. The name of the algorithm (e.g., "Simple", "TARP", "CeilingDiffuser", "AdaptiveConvectionAlgorithm").}
}
}
\usage{
l_idf_algorithm
}
\description{
This dataset contains a list of algorithms used in IDF (Input Data File) simulations.
Each algorithm is represented by a name.
}
\examples{
\dontrun{
# Access the first algorithm
l_idf_algorithm$algorithm[1]

# Access the second algorithm
l_idf_algorithm$algorithm[2]

# Access the third algorithm
l_idf_algorithm$algorithm[3]

# Access the fourth algorithm
l_idf_algorithm$algorithm[4]
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
