% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_idf_blocks}
\alias{l_idf_blocks}
\title{Dictionary with IDF commented blocks}
\format{
A list containing the following commented IDF blocks:
\itemize{
\item AirLoopHVAC: Represents the primary air loop in an HVAC system.
\item AirLoopHVAC:ControllerList: A list of controllers associated with an AirLoopHVAC.
\item AirLoopHVAC:OutdoorAirSystem: Defines the outdoor air system for an AirLoopHVAC.
\item AirLoopHVAC:OutdoorAirSystem:EquipmentList: A list of equipment in the outdoor air system.
\item AirLoopHVAC:ReturnPath: Defines the return air path in an AirLoopHVAC.
\item AirLoopHVAC:SupplyPath: Defines the supply air path in an AirLoopHVAC.
\item AirLoopHVAC:ZoneMixer: Represents a zone mixer in an AirLoopHVAC.
\item AirLoopHVAC:ZoneSplitter: Represents a zone splitter in an AirLoopHVAC.
\item AirTerminal:SingleDuct:ConstantVolume:NoReheat: A single duct air terminal with constant volume and no reheat.
\item AirflowNetwork:MultiZone:Surface: Defines a surface in a multi-zone airflow network.
\item AirflowNetwork:MultiZone:Zone: Defines a zone in a multi-zone airflow network.
\item AvailabilityManager:Scheduled: Manages the availability of equipment based on a schedule.
\item AvailabilityManagerAssignmentList: A list of availability managers assigned to equipment.
\item Boiler:HotWater: Represents a hot water boiler.
\item Branch: Defines a branch in a fluid loop.
\item BranchList: A list of branches in a fluid loop.
\item Building: Defines the building properties.
\item BuildingSurface:Detailed: Represents a detailed building surface.
\item Coil:Heating:Fuel: A heating coil that uses fuel.
\item Coil:WaterHeating:AirToWaterHeatPump:Pumped: A water heating coil using an air-to-water heat pump.
\item CondenserEquipmentList: A list of condenser equipment.
\item CondenserEquipmentOperationSchemes: Defines the operation schemes for condenser equipment.
\item CondenserLoop: Represents a condenser loop.
\item Connector:Mixer: A mixer connector.
\item Connector:Splitter: A splitter connector.
\item ConnectorList: A list of connectors.
\item Construction: Defines a construction element.
\item Construction:AirBoundary: Represents an air boundary construction.
\item ConstructionProperty:InternalHeatSource: Defines internal heat sources in a construction.
\item Controller:MechanicalVentilation: A controller for mechanical ventilation.
\item Controller:OutdoorAir: A controller for outdoor air.
\item ConvergenceLimits: Defines the convergence limits for simulations.
\item CurrencyType: Specifies the type of currency used.
\item Curve:Bicubic: A bicubic curve.
\item Curve:Biquadratic: A biquadratic curve.
\item Curve:Cubic: A cubic curve.
\item Curve:Exponent: An exponential curve.
\item Curve:Linear: A linear curve.
\item Curve:QuadLinear: A quad-linear curve.
\item Curve:Quadratic: A quadratic curve.
\item Curve:QuadraticLinear: A quadratic-linear curve.
\item Curve:Quartic: A quartic curve.
\item Curve:Triquadratic: A triquadratic curve.
\item DesignSpecification:OutdoorAir: Specifies the design for outdoor air.
\item DesignSpecification:ZoneAirDistribution: Specifies the design for zone air distribution.
\item DistrictCooling: Represents district cooling.
\item DistrictHeating:Water: Represents district heating using water.
\item ElectricEquipment: Defines electric equipment.
\item ElectricLoadCenter:Distribution: Represents the distribution centre for electric loads.
\item ElectricLoadCenter:Generators: A list of generators in the electric load centre.
\item ElectricLoadCenter:Inverter:Simple: A simple inverter in the electric load centre.
\item ElectricLoadCenter:Storage:Simple: Simple storage in the electric load centre.
\item Equipment: General equipment definition.
\item Exterior:FuelEquipment: Defines exterior fuel equipment.
\item Exterior:Lights: Defines exterior lighting.
\item Exterior:WaterEquipment: Defines exterior water equipment.
\item Fan:ConstantVolume: A constant volume fan.
\item Fan:OnOff: An on/off fan.
\item FenestrationSurface:Detailed: Represents a detailed fenestration surface.
\item GasEquipment: Defines gas equipment.
\item Generator:Photovoltaic: Represents a photovoltaic generator.
\item Generator:WindTurbine: Represents a wind turbine generator.
\item GlobalGeometryRules: Defines global geometry rules.
\item GroundHeatExchanger:ResponseFactors: Defines response factors for a ground heat exchanger.
\item GroundHeatExchanger:Slinky: Represents a slinky ground heat exchanger.
\item GroundHeatExchanger:System: Represents a ground heat exchanger system.
\item GroundHeatExchanger:Vertical:Properties: Defines properties for a vertical ground heat exchanger.
\item HVACTemplate:Thermostat: Represents an HVAC template for a thermostat.
\item HVACTemplate:Zone:IdealLoadsAirSystem: Represents an HVAC template for an ideal loads air system in a zone.
\item HeatBalanceAlgorithm: Defines the heat balance algorithm.
\item HeatExchanger:FluidToFluid: Represents a fluid-to-fluid heat exchanger.
\item HeatPump:WaterToWater:EquationFit:Heating: A water-to-water heat pump for heating using an equation fit model.
\item HeatPump:WaterToWater:ParameterEstimation:Heating: A water-to-water heat pump for heating using parameter estimation.
\item HotWaterEquipment: Defines hot water equipment.
\item Lights: Defines lighting.
\item Material: Defines a material.
\item Material:AirGap: Represents an air gap material.
\item Material:NoMass: Represents a no-mass material.
\item OtherEquipment: Defines other equipment.
\item OutdoorAir:Mixer: Represents an outdoor air mixer.
\item OutdoorAir:Node: Represents an outdoor air node.
\item OutdoorAir:NodeList: A list of outdoor air nodes.
\item Output:Constructions: Defines output for constructions.
\item Output:Diagnostics: Defines output for diagnostics.
\item Output:Meter: Defines output for meters.
\item Output:Surfaces:Drawing: Defines output for surface drawings.
\item Output:Surfaces:List: Defines output for surface lists.
\item Output:Table:SummaryReports: Defines output for summary reports in table format.
\item Output:Variable: Defines output variables.
\item Output:VariableDictionary: Defines the dictionary for output variables.
\item OutputControl:ReportingTolerances: Defines reporting tolerances for output control.
\item OutputControl:Table:Style: Defines the style for output tables.
\item People: Defines people in the simulation.
\item PhotovoltaicPerformance:Simple: Represents simple photovoltaic performance.
\item Pipe:Adiabatic: Represents an adiabatic pipe.
\item PlantEquipmentList: A list of plant equipment.
\item PlantEquipmentOperation:HeatingLoad: Defines operation for plant equipment based on heating load.
\item PlantEquipmentOperation:Uncontrolled: Defines uncontrolled operation for plant equipment.
\item PlantEquipmentOperationSchemes: Defines operation schemes for plant equipment.
\item PlantLoop: Represents a plant loop.
\item Pump:ConstantSpeed: A constant speed pump.
\item Pump:VariableSpeed: A variable speed pump.
\item RunPeriod: Defines the run period for the simulation.
\item Schedule:Compact: Represents a compact schedule.
\item Schedule:File: Represents a schedule from a file.
\item Schedule:Year: Represents a yearly schedule.
\item ScheduleTypeLimits: Defines limits for schedule types.
\item SetpointManager:FollowGroundTemperature: A setpoint manager that follows ground temperature.
\item SetpointManager:MixedAir: A setpoint manager for mixed air.
\item SetpointManager:Scheduled: A setpoint manager based on a schedule.
\item SetpointManager:SingleZone:Reheat: A setpoint manager for single zone reheat.
\item Shading:Building:Detailed: Represents detailed building shading.
\item ShadowCalculation: Defines shadow calculation parameters.
\item SimulationControl: Defines control parameters for the simulation.
\item Site:GroundReflectance: Defines ground reflectance for the site.
\item Site:GroundReflectance:SnowModifier: Defines snow modifier for ground reflectance.
\item Site:GroundTemperature:BuildingSurface: Defines ground temperature for building surface.
\item Site:GroundTemperature:Deep: Defines deep ground temperature.
\item Site:GroundTemperature:FCfactorMethod: Defines ground temperature using the FC factor method.
\item Site:GroundTemperature:Shallow: Defines shallow ground temperature.
\item Site:GroundTemperature:Undisturbed:KusudaAchenbach: Defines undisturbed ground temperature using the Kusuda-Achenbach method.
\item Site:Location: Defines the site location.
\item Site:WaterMainsTemperature: Defines water mains temperature for the site.
\item Sizing:Parameters: Defines sizing parameters.
\item Sizing:Plant: Defines plant sizing.
\item Sizing:System: Defines system sizing.
\item Sizing:Zone: Defines zone sizing.
\item SteamEquipment: Defines steam equipment.
\item SurfaceConvectionAlgorithm:Inside: Defines the inside surface convection algorithm.
\item SurfaceConvectionAlgorithm:Outside: Defines the outside surface convection algorithm.
\item ThermostatSetpoint:DualSetpoint: Represents a dual setpoint thermostat.
\item Timestep: Defines the simulation timestep.
\item Version: Defines the version of the simulation.
\item WaterHeater:HeatPump:PumpedCondenser: Represents a heat pump water heater with a pumped condenser.
\item WaterHeater:Mixed: Represents a mixed water heater.
\item WaterHeater:Sizing: Defines water heater sizing.
\item WaterHeater:Stratified: Represents a stratified water heater.
\item WaterUse:Connections: Defines water use connections.
\item WaterUse:Equipment: Defines water use equipment.
\item WindowProperty:FrameAndDivider: Defines window frame and divider properties.
\item Windowmaterial:Blind: Represents a window blind material.
\item Windowmaterial:Gas: Represents a window gas material.
\item Windowmaterial:Glazing: Represents a window glazing material.
\item Windowmaterial:Screen: Represents a window screen material.
\item Windowmaterial:Shade: Represents a window shade material.
\item Zone: Defines a zone in the building.
\item ZoneAirContaminantBalance: Defines the air contaminant balance for a zone.
\item ZoneAirHeatBalanceAlgorithm: Defines the air heat balance algorithm for a zone.
\item ZoneControl:Thermostat: Represents a thermostat for zone control.
\item ZoneHVAC:AirDistributionUnit: Represents an air distribution unit for zone HVAC.
\item ZoneHVAC:Baseboard:Convective:Electric: Represents an electric convective baseboard for zone HVAC.
\item ZoneHVAC:Baseboard:Convective:Water: Represents a water convective baseboard for zone HVAC.
\item ZoneHVAC:Baseboard:RadiantConvective:Electric: Represents an electric radiant convective baseboard for zone HVAC.
\item ZoneHVAC:Baseboard:RadiantConvective:Water: Represents a water radiant convective baseboard for zone HVAC.
\item ZoneHVAC:Baseboard:RadiantConvective:Water:Design: Represents the design for a water radiant convective baseboard for zone HVAC.
\item ZoneHVAC:EquipmentConnections: Defines equipment connections for zone HVAC.
\item ZoneHVAC:EquipmentList: A list of equipment for zone HVAC.
\item ZoneHVAC:IdealLoadsAirSystem: Represents an ideal loads air system for zone HVAC.
\item ZoneHVAC:LowTemperatureRadiant:VariableFlow: Represents a low temperature radiant system with variable flow for zone HVAC.
\item ZoneHVAC:LowTemperatureRadiant:VariableFlow:Design: Represents the design for a low temperature radiant system with variable flow for zone HVAC.
\item ZoneInfiltration:DesignFlowRate: Defines the design flow rate for zone infiltration.
\item ZoneInfiltration:EffectiveLeakageArea: Defines the effective leakage area for zone infiltration.
\item ZoneInfiltration:FlowCoefficient: Defines the flow coefficient for zone infiltration.
\item ZoneMixing: Defines zone mixing.
\item ZoneVentilation:DesignFlowRate: Defines the design flow rate for zone ventilation.
\item ZoneVentilation:WindandStackOpenArea: Defines the wind and stack open area for zone ventilation.
}
}
\usage{
l_idf_blocks
}
\description{
This object contains a dictionary with IDF commented blocks.
}
\examples{
\dontrun{
# Displaying the whole list
l_idf_blocks
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
