% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--layouts.R
\docType{data}
\name{l_room_lm}
\alias{l_room_lm}
\title{Linear Models for Room Layouts}
\format{
A list of 16 elements, each containing a linear model with the following components:
\describe{
\item{n_rooms}{A linear model for the number of rooms.}
\item{n_storeys}{A linear model for the number of storeys.}
\item{living_level}{A linear model for the living level.}
\item{living_function}{A linear model for the living function.}
\item{kitchen_level}{A linear model for the kitchen level.}
\item{kitchen_function}{A linear model for the kitchen function.}
\item{bedroom_level}{A linear model for the bedroom level.}
\item{bedroom_function}{A linear model for the bedroom function.}
\item{bathroom_level}{A linear model for the bathroom level.}
\item{extra_room_1 }{A linear model for the extra room 1.}
\item{extra_room_2}{A linear model for the extra room 2.}
\item{extra_room_3}{A linear model for the extra room 3.}
\item{extra_room_4}{A linear model for the extra room 4.}
\item{extra_room_5}{A linear model for the extra room 5.}
\item{extra_room_6}{A linear model for the extra room 6.}
\item{extra_room_7}{A linear model for the extra room 7.}
}
Each linear model contains the following components:
\describe{
\item{coefficients}{Named numeric vector of coefficients.}
\item{residuals}{Named numeric vector of residuals.}
\item{effects}{Named numeric vector of effects.}
\item{rank}{Integer rank of the model.}
\item{fitted.values}{Named numeric vector of fitted values.}
\item{assign}{Integer vector of variable assignments.}
\item{qr}{List containing QR decomposition components.}
\item{df.residual}{Integer degrees of freedom of residuals.}
\item{contrasts}{List of contrasts used in the model.}
\item{xlevels}{List of factor levels used in the model.}
\item{call}{The matched call.}
\item{terms}{The terms object used.}
\item{model}{The model frame.}
}
}
\usage{
l_room_lm
}
\description{
This list contains multiple linear models related to room layouts, including models for the number of rooms, number of storeys, living level, and living function.
}
\examples{
\dontrun{
# Access the linear model for the number of rooms
lm_n_rooms <- l_room_lm$n_rooms

# View the coefficients of the model
lm_n_rooms$coefficients

# Access the residuals of the model
lm_n_rooms$residuals

# Access the fitted values of the model
lm_n_rooms$fitted.values
}
}
\keyword{datasets}
