% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_sched_compact}
\alias{l_sched_compact}
\title{Dictionary with schedule information}
\format{
A list containing datasets related to IDF schedules.
\itemize{
\item schedule:thermostat-control-type: A list containing blocks of data related to thermostat control types.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:thermostat-control-type"
}
\item schedule:heating-availability-\{intermittent\}: A list containing blocks of data related to intermittent heating availability.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:heating-availability-\{intermittent\}"
}
\item schedule:heating-availability-\{constant\}: A list containing blocks of data related to constant heating availability.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:heating-availability-\{constant\}"
}
\item schedule:water-heater-ambient-temperature: A list containing blocks of data related to water heater ambient temperature.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:water-heater-ambient-temperature"
}
\item schedule:hot-water-services-warm: A list containing blocks of data related to warm hot water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:hot-water-services-warm"
}
\item schedule:hot-water-services-hot: A list containing blocks of data related to hot water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Any Number"
\item name: "schedule:hot-water-services-hot"
}
\item schedule:water-services:\{KitchenSink\}: A list containing blocks of data related to kitchen sink water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{KitchenSink\}"
}
\item schedule:water-services:\{KitchenOther\}: A list containing blocks of data related to other kitchen water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{KitchenOther\}"
}
\item schedule:water-services:\{KitchenWasher\}: A list containing blocks of data related to kitchen washer water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{KitchenWasher\}"
}
\item schedule:water-services:\{BathroomShower\}: A list containing blocks of data related to bathroom shower water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{BathroomShower\}"
}
\item schedule:water-services:\{BathroomSink\}: A list containing blocks of data related to bathroom sink water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{BathroomSink\}"
}
\item schedule:water-services:\{BathroomBath\}: A list containing blocks of data related to bathroom bath water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{BathroomBath\}"
}
\item schedule:water-services:\{BathroomService\}: A list containing blocks of data related to bathroom service water services.
\itemize{
\item blocks: A list containing specific blocks of data.
\item limits: "Fraction"
\item name: "schedule:water-services:\{BathroomService\}"
}
}
}
\usage{
l_sched_compact
}
\description{
This object contains a dictionary with schedule information.
}
\examples{
\dontrun{
# Displaying the whole list
l_sched_compact
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_limit}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
