% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--dictionaries.R
\docType{data}
\name{l_schedules_limit}
\alias{l_schedules_limit}
\title{Schedule Limits Data}
\format{
A list of 21 elements, each containing the following components:
\describe{
\item{values}{A list of values, where each value can be a character, integer, or numeric.}
}
}
\usage{
l_schedules_limit
}
\description{
This dataset contains various schedule limits used in simulations.
Each limit includes a set of values that define the constraints for the schedule.
}
\examples{
\dontrun{
# Access the first limit's values
l_schedules_limit[[1]]$values

# Access the values of the fourth limit
l_schedules_limit[[4]]$values

# Access the type of the sixth limit
l_schedules_limit[[6]]$values[[1]]
}
}
\seealso{
Other supporting dictionaries: 
\code{\link{l_activities}},
\code{\link{l_activity_groups}},
\code{\link{l_appliance_rating}},
\code{\link{l_comfort_states}},
\code{\link{l_energy_systems}},
\code{\link{l_epw_header_units}},
\code{\link{l_housing_groups}},
\code{\link{l_htg_curves}},
\code{\link{l_idf_algorithm}},
\code{\link{l_idf_blocks}},
\code{\link{l_idf_lzc}},
\code{\link{l_idf_outputs}},
\code{\link{l_sched_compact}},
\code{\link{l_schedules_airflow}},
\code{\link{l_schedules_base}},
\code{\link{l_schedules_people}}
}
\concept{supporting dictionaries}
\keyword{datasets}
