% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{list_evaluation_files}
\alias{list_evaluation_files}
\title{Lists files matching a pattern in a directory.}
\usage{
list_evaluation_files(relative_path, file_pattern)
}
\arguments{
\item{relative_path}{The relative path to the directory.}

\item{file_pattern}{The regular expression pattern for file names.}
}
\value{
A sorted vector of integer file IDs extracted from
the matched filenames.
}
\description{
This function retrieves a list of files
within a directory that match a specified regular
expression pattern.
}
