% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{locate_weather_files}
\alias{locate_weather_files}
\title{Locate and Download Weather Files}
\usage{
locate_weather_files(fname, l.par, d.epw = d_stations)
}
\arguments{
\item{fname}{Character string specifying the weather file name or source.}

\item{l.par}{List containing weather file source information.}

\item{d.epw}{Data frame containing weather station information.}
}
\value{
A list containing weather files and metadata.
}
\description{
Identifies and downloads weather files based on provided parameters and user preferences.
}
