% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{make_block_infiltration}
\alias{make_block_infiltration}
\title{Creates infiltration and ventilation data for a block}
\usage{
make_block_infiltration(d_ven, d_dwl, l_dms, i_vnt = TRUE)
}
\arguments{
\item{d_ven}{Ventilation data.}

\item{d_dwl}{Dwelling data.}

\item{l_dms}{Design meteorological data.}

\item{i_vnt}{Logical flag indicating whether to include ventilation.}
}
\value{
A list containing:
\itemize{
\item \code{idf}: A character vector of IDF lines.
\item \code{sum}: Summary data related to infiltration and ventilation rates.
}
}
\description{
Calculates infiltration and ventilation rates, generates IDF-compatible data,
and returns both the IDF data and summary information.
}
