% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{make_block_infiltration_leak}
\alias{make_block_infiltration_leak}
\title{Creates IDF blocks for infiltration leak rates}
\usage{
make_block_infiltration_leak(d_rate, l_sup = l_idf_blocks)
}
\arguments{
\item{d_rate}{Dataframe containing ventilation rates.}

\item{l_sup}{List of supplementary information, typically IDF block templates.}
}
\value{
A character vector containing IDF lines for ventilation rates.
}
\description{
Generates IDF blocks for infiltration leak rates based on provided data.
}
