% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{make_block_infiltration_rate}
\alias{make_block_infiltration_rate}
\title{Creates IDF blocks for infiltration rates}
\usage{
make_block_infiltration_rate(d_rate, l_sup = l_idf_blocks)
}
\arguments{
\item{d_rate}{Dataframe containing infiltration rates.}

\item{l_sup}{List of supplementary information, typically IDF block templates.}
}
\value{
A character vector containing IDF lines for infiltration rates.
}
\description{
Generates IDF blocks for infiltration rates based on provided data.
}
