% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_block_output_general}
\alias{make_block_output_general}
\title{Creates a General Block Output}
\usage{
make_block_output_general(input_parameters = l_idf_blocks)
}
\arguments{
\item{input_parameters}{A list of input parameters. Defaults to \code{l_idf_blocks}.}
}
\value{
A list of output parameters.
}
\description{
Generates a list of output parameters based on the provided \code{input_parameters} list.
}
