% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_block_output_metres}
\alias{make_block_output_metres}
\title{Creates Output Meters for an IDF Block}
\usage{
make_block_output_metres(
  label_list,
  heating_data,
  load_zone_calculations,
  resolution_key,
  support_list = l_idf_blocks
)
}
\arguments{
\item{label_list}{A list of available outcome variables.}

\item{heating_data}{A list of heating system data.}

\item{load_zone_calculations}{A list of load zone calculations.}

\item{resolution_key}{A string indicating the resolution of the output.}

\item{support_list}{A support list. Defaults to \code{l_idf_blocks}.}
}
\value{
A character vector representing the IDF block structure.
}
\description{
This function generates a list of output meters based on provided parameters.
}
