% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_block_output_variables}
\alias{make_block_output_variables}
\title{Creates Output Variables for an IDF Block}
\usage{
make_block_output_variables(
  label_list,
  load_zone_calculations,
  system_key,
  result_key,
  support_list = l_idf_blocks
)
}
\arguments{
\item{label_list}{A label list.}

\item{load_zone_calculations}{A list of load zone calculations.}

\item{system_key}{A system key.}

\item{result_key}{A result key.}

\item{support_list}{A support list. Defaults to \code{l_idf_blocks}.}
}
\value{
A character vector representing the IDF block structure.
}
\description{
This function generates a list of output variables based on provided parameters.
}
