% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_block_output_zonal}
\alias{make_block_output_zonal}
\title{Creates Output Variables for Zonal Data in an IDF Block}
\usage{
make_block_output_zonal(
  label_list,
  zones,
  result_key,
  support_list = l_idf_blocks
)
}
\arguments{
\item{label_list}{A label list.}

\item{zones}{A list of zones.}

\item{result_key}{A result key.}

\item{support_list}{A support list. Defaults to \code{l_idf_blocks}.}
}
\value{
A character vector representing the IDF block structure.
}
\description{
This function generates a list of output variables for zonal data based on provided parameters.
}
