% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{make_block_schedule}
\alias{make_block_schedule}
\title{Creates an IDF Block Schedule}
\usage{
make_block_schedule(l_zon, d_ref, l_sup = l_idf_blocks)
}
\arguments{
\item{l_zon}{A list of zones.}

\item{d_ref}{Reference data for block creation.}

\item{l_sup}{Supplementary information (default: l_idf_blocks).}

\item{iz}{Zone identifier.}

\item{block}{Block type.}

\item{ipar}{Input parameters.}

\item{com}{Comments for the block.}
}
\value{
A list containing the constructed IDF block schedule.

A list representing the IDF block.
}
\description{
Constructs an IDF block schedule based on provided zones, reference data, and supplementary information.

Constructs an individual IDF block with specified parameters.
}
