% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{make_block_ventilation_stack}
\alias{make_block_ventilation_stack}
\title{Creates IDF blocks for wind and stack ventilation}
\usage{
make_block_ventilation_stack(d_rate, l_sup = l_idf_blocks)
}
\arguments{
\item{d_rate}{Dataframe containing ventilation rates.}

\item{l_sup}{List of supplementary information, typically IDF block templates.}
}
\value{
A character vector containing IDF lines for wind and stack ventilation.
}
\description{
Generates IDF blocks for wind and stack ventilation
based on provided data.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
