% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_block_water_connections}
\alias{make_block_water_connections}
\title{Creates IDF blocks for water connections.}
\usage{
make_block_water_connections(d_ser, d_nodes, l_sup = l_idf_blocks)
}
\arguments{
\item{d_ser}{A data frame containing service information.}

\item{d_nodes}{A data frame containing node information.}

\item{l_sup}{A list containing IDF block data (optional).}
}
\value{
A character vector representing the IDF block content.
}
\description{
This function generates IDF blocks for water connections based on service and node data.
}
