% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_block_water_mains}
\alias{make_block_water_mains}
\title{Creates an IDF block for water mains temperature.}
\usage{
make_block_water_mains(l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup}{A list containing IDF block data.}
}
\value{
A character vector representing the IDF block.
}
\description{
This function generates an IDF block for water mains temperature based on provided data.
}
