% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{make_construction_glazing}
\alias{make_construction_glazing}
\title{Create IDF block for Windowmaterial:Glazing objects based on
list of materials}
\usage{
make_construction_glazing(l_iter, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing materials}

\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing glazing
objects using an iterator (\code{l_iter}) defining list of materials. It
also uses supplementary data (\code{l_sup}) for comments.
}
