% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{make_construction_layers}
\alias{make_construction_layers}
\title{Create IDF block for Construction objects based on list of layers}
\usage{
make_construction_layers(l_iter, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing list of layers.}

\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a
construction object an iterator (\code{l_iter}) defining list of layers.
It also uses supplementary data (\code{l_sup}) for comments.
}
