% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_condenser_schemes}
\alias{make_energy_block_condenser_schemes}
\title{Create IDF block for an object representing a
CondenserEquipmentOperationSchemes object}
\usage{
make_energy_block_condenser_schemes(l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a condenser
operation scheme using default properties. It also uses
supplementary data (\code{l_sup}) for comments.
}
