% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_ground_temperature_ka}
\alias{make_energy_block_ground_temperature_ka}
\title{Create IDF block for an object representing a
Site:GroundTemperature:Undisturbed:KusudaAchenbach object}
\usage{
make_energy_block_ground_temperature_ka(l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a ground
temperature KusudaAchenbach using default properties. It also uses
supplementary data (\code{l_sup}) for comments.
}
