% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_setpoint_manager}
\alias{make_energy_block_setpoint_manager}
\title{Create IDF block for Setpoint Manager based on loop data}
\usage{
make_energy_block_setpoint_manager(l_iter, d_loop, l_sup = l_idf_blocks)
}
\arguments{
\item{l_iter:}{List containing loop properties (e.g., name, temperature).}

\item{d_loop:}{Data frame containing detailed loop information.}

\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}

\item{i:}{Character string representing the loop type}

\item{d_lp:}{Data frame containing loop information (groups, types, ...)}

\item{i}{A character string specifying the node name.}

\item{li}{A list containing data for different nodes.}

\item{ln}{A list containing node information.}
}
\value{
A character vector representing the complete IDF block content.

A list containing information about plant inlet/outlet,
demand inlet/outlet, branches, and connectors for the  loop type.
Assigns spatial nodes to data.

This function assigns spatial nodes to data based on given parameters.

A data frame with columns \code{name}, \code{fluid}, and \code{node_sp}.
}
\description{
This function generates an IDF block representing a setpoint
manager using loop data (\code{d_loop}) and an iterator (\code{l_iter}) defining loop
properties. It also uses supplementary data (\code{l_sup}) for comments.

This function extracts information about nodes-connections
for a specific loop type (\code{i}) from a loop data frame (\code{d_lp}).
}
