% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_underfloor_radiator_design}
\alias{make_energy_block_underfloor_radiator_design}
\title{Create IDF block for ZoneHVAC:LowTemperatureRadiant:VariableFlow:Design
objects based on loop data}
\usage{
make_energy_block_underfloor_radiator_design(l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a zone
emitter design object using loop data (\code{d_loop}) and an iterator
(\code{l_iter}) defining loop properties. It also uses supplementary
data (\code{l_sup}) for comments.
}
