% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_warm_air_coil}
\alias{make_energy_block_warm_air_coil}
\title{Create IDF block for an object representing a Coil:Heating:Fuel
object}
\usage{
make_energy_block_warm_air_coil(d_item, l_sup = l_idf_blocks)
}
\arguments{
\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}

\item{d_item:}{Data frame containing system loop information.}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing a coil
component using system info (\code{d_item}) and default properties. It
also uses supplementary data (\code{l_sup}) for comments.
}
