% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{make_energy_block_warm_air_out_air_loop_assignment_manager}
\alias{make_energy_block_warm_air_out_air_loop_assignment_manager}
\title{Create IDF block for a AvailabilityManagerAssignmentList object
based on loop data}
\usage{
make_energy_block_warm_air_out_air_loop_assignment_manager(
  d_zone,
  l_sup = l_idf_blocks
)
}
\arguments{
\item{l_sup:}{List with supplementary data (defaults to \code{l_idf_blocks}).}
}
\value{
A character vector representing the complete IDF block content.
}
\description{
This function generates an IDF block representing an
availability manager assignment list object using default parameters.
It also  uses supplementary data (\code{l_sup}) for comments.
}
