% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-format.R
\name{make_idf_block_header}
\alias{make_idf_block_header}
\title{Create IDF Block Header}
\usage{
make_idf_block_header(txt, typ = "class", k_max = 65)
}
\arguments{
\item{txt}{The text for the header.}

\item{typ}{The type of header (default: 'class').}

\item{k_max}{The maximum length of the header (default: 65).}
}
\value{
A character string representing the IDF block header.
}
\description{
This function generates a formatted header for an IDF (Input Data File) block.
}
\examples{
# Generate a header for a class:
make_idf_block_header("Example Header")

# Generate a header using an alternative format:
make_idf_block_header("Example Header", typ = 'other', k_max = 70)
}
\seealso{
Other idf-formatting: 
\code{\link{add_idf_comment}()},
\code{\link{add_idf_comment_custom}()},
\code{\link{determine_end_of_idf_block}()},
\code{\link{iterate_on_list}()},
\code{\link{iterate_on_output_list}()},
\code{\link{make_idf_block}},
\code{\link{make_idf_block_title}()}
}
\concept{idf-formatting}
