% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-format.R
\name{make_idf_block_title}
\alias{make_idf_block_title}
\title{Create IDF Block Title}
\usage{
make_idf_block_title(lbl)
}
\arguments{
\item{lbl}{The label for the title.}
}
\value{
A character string representing the IDF block title.
}
\description{
This function creates a formatted title for an IDF (Input Data File) block.
}
\examples{
# Generate a title for a class:
make_idf_block_title("Example Class")
}
\seealso{
Other idf-formatting: 
\code{\link{add_idf_comment}()},
\code{\link{add_idf_comment_custom}()},
\code{\link{determine_end_of_idf_block}()},
\code{\link{iterate_on_list}()},
\code{\link{iterate_on_output_list}()},
\code{\link{make_idf_block}},
\code{\link{make_idf_block_header}()}
}
\concept{idf-formatting}
