% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{make_sim_period}
\alias{make_sim_period}
\title{Creates Simulation Period Data Frame}
\usage{
make_sim_period(yr)
}
\arguments{
\item{yr}{The simulation year}
}
\value{
A list containing two data frames: \code{perMonth} with days per
month and \code{period} with hourly time stamps.
}
\description{
Generates a data frame containing hourly time stamps
for a specified year.
}
\details{
This function creates a data frame with hourly time stamps for a
specified year. It also calculates days per month for the year.
The function returns a list containing these two data frames.
}
