% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{mirror_layout}
\alias{mirror_layout}
\title{Generate Mirrored or Equal Cuboids for Attachments}
\usage{
mirror_layout(l_it, b_att, b_sto, m_dim, l_drn, l_inf)
}
\arguments{
\item{l_it}{A list of room layouts.}

\item{b_att}{A string indicating the attachment condition ('left', 'right', 'both', or other).}

\item{b_sto}{A vector of storeys/floors.}

\item{m_dim}{A matrix of dimensions.}

\item{l_drn}{A list of dimensions for each room.}

\item{l_inf}{A list of custom input parameters.}
}
\value{
A list containing the mirrored layout for the specified attachment condition.
}
\description{
The function processes the input layout to generate mirrored layouts for the left and right sides. It then adjusts the origins of the new layouts and combines them based on the specified attachment condition.
}
\examples{
\dontrun{
  result <- mirror_layout(l_it, b_att, b_sto, m_dim, l_drn, l_inf)
  print(result)
}
Helper function to mirror a single layout
}
