% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_annual_temperature}
\alias{obtain_annual_temperature}
\title{Calculates annual temperature summaries}
\usage{
obtain_annual_temperature(df)
}
\arguments{
\item{df}{A data frame containing temperature data.}
}
\value{
A list of data frames containing three elements:
\itemize{
\item \code{monthly}: Monthly mean temperature for each month.
\item \code{hourly}: Hourly mean temperature for each month and zone.
\item \code{on_heat}: Mean temperature for each month and zone, grouped by heating status.
}
}
\description{
This function calculates annual temperature summaries based on the provided data frame.
}
