% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_archetype_dimensions}
\alias{obtain_archetype_dimensions}
\title{Obtain archetype dimensions}
\usage{
obtain_archetype_dimensions(idc, df_ref)
}
\arguments{
\item{idc}{A character vector specifying the archetype code.}

\item{df_ref}{A data frame containing complementary and summarised data.}
}
\value{
A data frame containing archetype dimensions.
}
\description{
This function extracts archetype dimensions from
a reference data frame.
}
