% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_archetype_level}
\alias{obtain_archetype_level}
\title{Obtain archetype level information}
\usage{
obtain_archetype_level(
  num_floors,
  basement_presence,
  attic_presence,
  reference_data
)
}
\arguments{
\item{num_floors}{An integer specifying the number of floors in the archetype.}

\item{basement_presence}{A character indicating the presence of a basement}

\item{attic_presence}{A character indicating the presence of an attic}

\item{reference_data}{A data frame containing a reference floor position variable}
}
\value{
A list containing information about the effective storey levels.
}
\description{
This function determines the effective storey levels
for an archetype based on the presence of a basement, attic,
and number of floors.
}
