% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_archetype_parameters}
\alias{obtain_archetype_parameters}
\title{Obtains archetype parameters from reference data.}
\usage{
obtain_archetype_parameters(housing_code, reference_data)
}
\arguments{
\item{housing_code}{Housing code.}

\item{reference_data}{Reference data containing complementary and
summarized information.}
}
\value{
A data frame with extracted archetype parameters.
}
\description{
This function extracts relevant parameters from the provided
reference data based on the housing code. Parse attachment conditions:
\enumerate{
\item{two attachments    ☐ ☐ ☐}
\item{left attachments    ☐ ☐ -}
\item{right attachments    - ☐ ☐}
\item{right attachments   - ☐ - }
}
}
