% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_building_zones}
\alias{obtain_building_zones}
\title{Generate IDF door objects for each of the surfaces}
\usage{
obtain_building_zones(l_lay)
}
\arguments{
\item{l_lay}{list of layout information}
}
\value{
list of thermal zones and spaces
}
\description{
Generate IDF door objects for each of the surfaces
}
\examples{
\dontrun{
obtain_building_zones(l_room_layouts)
}
}
