% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_carbon_and_primary}
\alias{obtain_carbon_and_primary}
\title{Calculates Carbon Emissions and Primary Energy Consumption}
\usage{
obtain_carbon_and_primary(df, df_ref = s_sap$CostAndCoefficientsSummary)
}
\arguments{
\item{df}{A data frame containing energy use data with columns for energy__kWh and fuel type.}

\item{df_ref}{(optional) A data frame containing reference coefficients for CO2 emissions and primary energy factors. Defaults to \code{s_sap$CostAndCoefficientsSummary}.}
}
\value{
A data frame with additional columns for CO2 emissions (CO2.eq) and primary energy consumption (primary).
}
\description{
Calculates carbon emissions and primary energy consumption based on energy use
data and reference coefficients.
}
