% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_comfort_parameters}
\alias{obtain_comfort_parameters}
\title{Obtains comfort parameters from building simulation data}
\usage{
obtain_comfort_parameters(df)
}
\arguments{
\item{df}{A data frame containing building simulation data.}
}
\value{
A list containing two data frames:
\itemize{
\item \code{pmv}: A data frame summarizing the distribution of PMV values.
\item \code{comfort_temperature}: A data frame summarizing comfort states based on internal temperature.
}
}
\description{
Estimates preliminary comfort states based on air temperature and PMV values.
}
\details{
This function processes data related to zone air temperature and PMV (Predicted
Mean Vote) to provide insights into occupant comfort. It calculates average
internal temperature and classifies it as "a bit too hot," "ok," or "a bit too cold"
based on predefined comfort limits. It also analyses PMV values and categorizes
them into comfort ranges using defined labels.
}
