% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_cross_indices}
\alias{obtain_cross_indices}
\title{Creates cross-indices for TUS and EHS datasets.}
\usage{
obtain_cross_indices(
  d_tus = s_tus$tbl_individual,
  d_ehs = tbl.EHS.common,
  path = path_TUS
)
}
\arguments{
\item{d_tus}{A data frame containing individual-level data (default:
s_tus$tbl_individual).}

\item{d_ehs}{A data frame containing household-level data (default:
tbl.EHS.common).}

\item{path}{The path to the output directory (default: path_TUS).}
}
\value{
A list containing the processed datasets (d_ehs and d_tus).
}
\description{
This function processes two datasets (d_tus and d_ehs),
creates cross-indices for various variables, and saves the results as CSV
files.
}
