% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewables.R
\name{obtain_daily_lzc}
\alias{obtain_daily_lzc}
\title{Calculates daily solar energy potential for a specific dwelling.}
\usage{
obtain_daily_lzc(
  idc,
  s_epw,
  s_ehs_ref = s_ehs_2011_ext,
  s_sap_ref = s_sap,
  s_bredem_ref = s_bredem
)
}
\arguments{
\item{idc}{Character string representing the dwelling code.}

\item{s_epw}{A list containing weather data (assumed to include latitude information).}

\item{s_ehs_2011_ext}{A list containing dwelling and household information (assumed to include geometry and complementary data).}

\item{s_sap}{A list containing solar access pattern data (assumed to include solar declination data).}

\item{s_bredem}{A list containing BRE DEM data (assumed to include flux constants and global irradiance data).}
}
\value{
A numeric value representing the daily solar energy potential.
}
\description{
This function calculates the daily solar energy potential for a dwelling based on weather data and building geometry.
}
