% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{obtain_days_per_month}
\alias{obtain_days_per_month}
\title{Obtain number of days in a month}
\usage{
obtain_days_per_month(month, year = NULL)
}
\arguments{
\item{month}{An integer or character representing the month (1-12 or month name).}

\item{year}{An optional integer representing the year (defaults to current year).}
}
\value{
A data frame with two columns: "Month" (character abbreviation) and "Days" (integer).
}
\description{
This function calculates the number of days in a given month.
}
\examples{
\dontrun{
obtain_days_per_month(5)  # Get days in May
}

}
