% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{obtain_ddy}
\alias{obtain_ddy}
\title{Extract Design Day Objects from IDF File}
\usage{
obtain_ddy(idf_file, rain_value, design_day_family = "heating")
}
\arguments{
\item{idf_file}{A character vector containing the lines of the IDF file.}

\item{rain_value}{The rain value to be included in Design Day objects.}

\item{design_day_family}{The type of Design Day objects to extract ('heating', 'cooling', or 'other'). Defaults to 'heating'.}
}
\value{
A list containing:
\itemize{
\item \code{idf}: A character vector with the modified IDF file content.
\item \code{sum}: A data frame summarizing the extracted Design Day objects (label and group).
}
}
\description{
This function parses an IDF file (\code{idf_file}) and extracts information
about Design Day objects based on the specified family (\code{design_day_family}).
It modifies the IDF file to include a rain value (\code{rain_value}) for those objects.
}
