% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_ddy_summary}
\alias{obtain_ddy_summary}
\title{Extracts Day/Year Summary Information}
\usage{
obtain_ddy_summary(df, yr)
}
\arguments{
\item{df}{A data frame containing day/year summary data.}

\item{yr}{The year of the data.}
}
\value{
A data frame containing day/year summary information.
}
\description{
Extracts day/year summary information from a given data frame.
}
