% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_design_days}
\alias{obtain_design_days}
\title{Assigns labels to data frame rows based on time ranges}
\usage{
obtain_design_days(df, df_index)
}
\arguments{
\item{df}{A data frame to which labels will be assigned.}

\item{df_index}{A data frame containing time ranges and corresponding labels.}
}
\value{
The input data frame with an added \code{ddy_item} column containing labels.
}
\description{
This function assigns labels to rows of a data frame based on time ranges
defined in another data frame.
}
