% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_dimensions}
\alias{obtain_dimensions}
\title{Obtain detailed list of geometrical parameters of house}
\usage{
obtain_dimensions(b_cub, rir_h = 1.125, rtv_h = 0.25, l_template = t_layouts)
}
\arguments{
\item{b_cub}{list of geometrical input-parameters}

\item{rir_h}{room-in-roof height}

\item{rtv_h}{top-roof/void height}

\item{l_template}{tuple of tupple of available layouts}
}
\value{
list of detailed geometrical parameters
}
\description{
Obtain detailed list of geometrical parameters of house
}
