% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_employment_bands}
\alias{obtain_employment_bands}
\title{Classify employment status into bands}
\usage{
obtain_employment_bands(employment_status)
}
\arguments{
\item{employment_status}{A character vector containing employment status
information.}
}
\value{
A character vector of the same length as \code{employment_status} with
employment bands.
}
\description{
Classifies employment status information (\code{employment_status})
into bands.
}
