% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_end_use_fuel}
\alias{obtain_end_use_fuel}
\title{Extracts End Use Fuel Data}
\usage{
obtain_end_use_fuel(df)
}
\arguments{
\item{df}{A data frame containing end use fuel data.}
}
\value{
A data frame containing end use fuel data with cleaned and processed columns.
}
\description{
Extracts end use fuel data from a given data frame.
}
