% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_end_use_summary}
\alias{obtain_end_use_summary}
\title{Calculates End-Use Energy Summary}
\usage{
obtain_end_use_summary(df, i_tfa, df_sum)
}
\arguments{
\item{df}{A data frame containing energy use data.}

\item{i_tfa}{The total floor area.}

\item{df_sum}{A data frame containing summary data.}
}
\value{
A data frame summarizing end-use energy consumption and intensity.
}
\description{
Calculates energy consumption and intensity for various end-use categories
from a data frame and summary data.
}
