% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_energy_demand_profile}
\alias{obtain_energy_demand_profile}
\title{Converts energy data to kWh and returns JSON}
\usage{
obtain_energy_demand_profile(data)
}
\arguments{
\item{data}{A data frame containing energy data.}
}
\value{
A JSON string representing the energy data in kWh.
}
\description{
This function takes a data frame containing energy data in Joules, converts the
energy values to kilowatt-hours (kWh), renames columns accordingly, and returns
the data as a JSON string.
}
