% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_energy_system_performance}
\alias{obtain_energy_system_performance}
\title{Extracts energy system performance data based on system code}
\usage{
obtain_energy_system_performance(df, l_sum)
}
\arguments{
\item{df}{A data frame containing energy system data.}

\item{l_sum}{A list containing design day information and potentially
the system code (may need adjustment based on your data structure).}
}
\value{
A list containing extracted performance data or '-' if the system code is not recognized.
}
\description{
This function extracts relevant data for energy system performance
based on the system code provided in the summary data.
}
