% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_fenestration_summary}
\alias{obtain_fenestration_summary}
\title{Obtain summary data of fenestration objects}
\usage{
obtain_fenestration_summary(l_gl)
}
\arguments{
\item{l_gl}{list of generated glazing/door objects}
}
\value{
a tibble summary
}
\description{
Obtain summary data of fenestration objects
}
\examples{
\dontrun{
obtain_fenestration_summary(l_glazing)
}
}
