% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_flow_temperatures}
\alias{obtain_flow_temperatures}
\title{Extracts flow temperatures from data frame}
\usage{
obtain_flow_temperatures(df, l_sum)
}
\arguments{
\item{df}{A data frame containing temperature data.}

\item{l_sum}{A list containing system information (including system code).}
}
\value{
A list containing extracted temperature data.
}
\description{
This function extracts inlet and outlet temperatures for different system components
based on the provided system code.
}
