% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_fuel_demand}
\alias{obtain_fuel_demand}
\title{Extracts Fuel Demand Information}
\usage{
obtain_fuel_demand(df, i_tfa)
}
\arguments{
\item{df}{A data frame containing fuel demand data.}

\item{i_tfa}{The total floor area.}
}
\value{
A list containing two data frames:
\itemize{
\item \code{table}: Detailed fuel demand data.
\item \code{summary}: Summary of fuel demand and energy intensity.
}
}
\description{
Extracts fuel demand data and calculates energy intensity from a data frame.
}
