% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_glazing_areas}
\alias{obtain_glazing_areas}
\title{Obtain glazing dimensions across the house}
\usage{
obtain_glazing_areas(l_lay, l_dmg, l_cbd)
}
\arguments{
\item{l_lay}{list of existing room layouts}

\item{l_dmg}{list of major dimensions}

\item{l_cbd}{list of basic geometrical parameters}
}
\value{
list of glazing areas
}
\description{
Obtain glazing dimensions across the house
}
