% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_glazing_contribution}
\alias{obtain_glazing_contribution}
\title{Calculates hourly glazing contribution to heating}
\usage{
obtain_glazing_contribution(df)
}
\arguments{
\item{df}{A data frame containing date/time, heating status, and window
heat loss/gain data.}
}
\value{
A data frame summarizing average glazing contribution for each hour.
}
\description{
This function calculates the hourly average contribution of glazing
(windows) to heating by analysing heat loss and gain rates.
}
