% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_glazing_dimensions}
\alias{obtain_glazing_dimensions}
\title{Obtain glazing dimensions across the house}
\usage{
obtain_glazing_dimensions(i_rly, l_rly, l_gla, k_main = TRUE)
}
\arguments{
\item{i_rly}{list of storeys}

\item{l_rly}{list of room layouts}

\item{l_gla}{list of glazing parameters}

\item{k_main}{indication of building-unit}
}
\value{
list of estimated glazing dimensions
}
\description{
Obtain glazing dimensions across the house
}
