% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_hdd_profile}
\alias{obtain_hdd_profile}
\title{Calculates Heating Degree Days (HDD) Profile}
\usage{
obtain_hdd_profile(df, k)
}
\arguments{
\item{df}{A data frame containing weather data with columns \code{Date.Time} and
\code{Outdoor.Air.Drybulb.Temperature..C..Hourly.}.}

\item{k}{The base temperature for calculating HDD.}
}
\value{
A list containing:
\itemize{
\item \code{table}: A data frame with monthly mean temperature and total HDD.
\item \code{hdd}: A list with \code{threshold} (base temperature) and \code{total} HDD.
}
}
\description{
This function calculates the HDD profile for a given dataset and base temperature (k).
}
