% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_heat_transfer}
\alias{obtain_heat_transfer}
\title{Combines envelope information and heat storage data}
\usage{
obtain_heat_transfer(dp, tx)
}
\arguments{
\item{dp}{Data frame containing building envelope or heat storage information.}

\item{tx}{Data frame containing envelope information.}
}
\value{
A list containing envelope information and heat storage rate data.
}
\description{
This function extracts envelope information and heat storage rate from
provided data sources and combines them into a list.
}
