% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_heat_transfer_ddy}
\alias{obtain_heat_transfer_ddy}
\title{Calculates heat transfer by design day item}
\usage{
obtain_heat_transfer_ddy(dp)
}
\arguments{
\item{dp}{A data frame containing heat transfer data with a \code{ddy_item} column.}
}
\value{
A list of data frames, each representing heat transfer data for a specific \code{ddy_item}.
}
\description{
This function groups the input data by \code{ddy_item} and calculates the inside heat transfer rate for each group.
}
