% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-system.R
\name{obtain_heating_components}
\alias{obtain_heating_components}
\title{Obtain heating components based on energy system ID}
\usage{
obtain_heating_components(id, ins, l_ref = l_energy_systems)
}
\arguments{
\item{id:}{The energy system ID}

\item{ins:}{Additional insulation parameters}

\item{l_ref:}{A reference list containing energy system
data (default: l_energy_systems).}
}
\value{
A list containing loop type, emitter type, space heating,
and hot water information.
}
\description{
This function extracts heating components based on the
provided energy system ID.
}
\examples{
\dontrun{
id=k_heating_code
ins=k_insulation
l_ref=l_energy_systems
obtain_heating_components(id, ins, l_ref)
}
}
