% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_heating_efficiencies}
\alias{obtain_heating_efficiencies}
\title{Calculates Heating Efficiencies}
\usage{
obtain_heating_efficiencies(l_sum, l_perf, m_usge, df_sum, l_idf, l_heat)
}
\arguments{
\item{l_sum}{A list containing summary information.}

\item{l_perf}{A list containing performance data.}

\item{m_usge}{A matrix containing energy usage data.}

\item{df_sum}{A data frame containing summary data.}

\item{l_idf}{A list containing IDF data.}

\item{l_heat}{A list for storing heating efficiency results.}
}
\value{
A list containing calculated heating efficiencies.
}
\description{
Calculates heating system efficiencies based on provided data.
}
