% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_heating_size}
\alias{obtain_heating_size}
\title{Extracts Heating System Information}
\usage{
obtain_heating_size(df, l_sum)
}
\arguments{
\item{df}{A data frame containing system information data.}

\item{l_sum}{A list containing a summary of the house energy system.}
}
\value{
A list containing extracted information:
\itemize{
\item report: Central plant information (or "not applicable").
\item heater: Water heater information (or "not applicable").
\item heat_pump: Heat pump water heater information (or "not applicable").
\item coil: Coil sizing information (or "not applicable").
\item sizing: Zone sizing information.
\item components: Component sizing information.
}
}
\description{
Extracts information on central plant, water heater, heat pump water heater,
coil sizing, zone sizing, and component sizing from a data frame and a summary list.
}
