% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_heating_status}
\alias{obtain_heating_status}
\title{Determines heating status based on data frame content}
\usage{
obtain_heating_status(df)
}
\arguments{
\item{df}{A data frame containing columns related to heating system operation.}
}
\description{
This function analyses a data frame containing various heating-related
columns and infers the heating system status (on/off) and heating rate.
}
