% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_hlp_building}
\alias{obtain_hlp_building}
\title{Obtain building thermal properties}
\usage{
obtain_hlp_building(tblFlrs, tblOpns, tblOpqs, tblInts, varAge, varTFA)
}
\arguments{
\item{tblFlrs}{Data table containing floor information (potentially unused)}

\item{tblOpns}{Data table containing information about openings (windows etc.)}

\item{tblOpqs}{Data table containing information about opaque elements (walls etc.)}

\item{tblInts}{Data table containing information about internal elements}

\item{varAge}{Numeric variable representing building age}

\item{varTFA}{Numeric variable representing the total floor area}
}
\value{
Data frame containing thermal properties: opening ratio,
fabric heat loss, heat capacity, thermal mass, thermal bridge,
and total heat loss.
}
\description{
This function calculates various thermal properties
of a building based on data tables and additional variables.
}
