% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_hlp_floors_bbgf}
\alias{obtain_hlp_floors_bbgf}
\title{Obtain U-values for basement and ground floors}
\usage{
obtain_hlp_floors_bbgf(idc, l_inp)
}
\arguments{
\item{idc}{A character string specifying the archetype code.}

\item{l_inp}{A list containing three data frames:
\itemize{
\item The first data frame contains information about the
building geometry (D030_WallThickness, D059_BasementFloorHeatLossArea,
D113_BasementFloorExposedPerimeter, D060_GroundFloorHeatLossArea,
D114_GroundFloorExposedPerimeter).
\item The second data frame contains information about the
floor construction (D074_BasementFloorConstructionType,
D075_GroundFloorConstructionType).
\item The third data frame contains information about the
reference region (D003_Region).
}}
}
\value{
A data frame containing U-values for basement and ground floors
(\code{BB} and \code{GF} columns) along with other calculated parameters.
}
\description{
This function calculates the U-values (thermal transmittance)
for basement and ground floors in a building based on various construction
parameters and reference data. Based on SAP S5.4 U-values of floors
next to the ground
}
