% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_hlp_internals}
\alias{obtain_hlp_internals}
\title{Obtain internal material properties}
\usage{
obtain_hlp_internals(idc, l_inp)
}
\arguments{
\item{idc}{Housing code}

\item{l_inp}{List containing data tables:
\itemize{
\item Element 1: Gmt table
\item Element 2: Lss table
\item Element 3: Cmp table (unused)
\item Element 4: Inf table
\item Element 5: Usr table (unused)
}}
}
\value{
Data frame with internal material properties for party wall, party
floor, party ceiling, internal wall, internal floor, and internal ceiling.
}
\description{
This function retrieves internal material properties
from various data sources based on a housing code.
}
