% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_hlp_opaques}
\alias{obtain_hlp_opaques}
\title{Obtain heat loss parameters for opaque building elements}
\usage{
obtain_hlp_opaques(idc, l_inp)
}
\arguments{
\item{idc}{A character string specifying the archetype code.}

\item{l_inp}{A list containing five data frames:
\itemize{
\item The first data frame contains information about the building geometry
(D036_GroundFloorArea, D040_RoomInRoofArea, D059_BasementFloorHeatLossArea,
D061_RoofArea, D064_BasementWallArea, D065_ExternalWallArea,
D121_SemiExposedArea).
\item The second data frame contains information about the construction types
of specific elements (D074_BasementFloorConstructionType,
D075_GroundFloorConstructionType, D076_BasementWallConstruction,
D077_ExternalWallConstruction, D079_RoofConstruction,
D080_LoftInsulationThickness, D081_RoomInRoofConstruction,
H058_ExposedFloorConstruction, H059_ExposedFloorHeatLossArea).
\item The third data frame contains information about user-defined U-values
(Utotal).
\item The fourth data frame contains information on the dwelling age
(D001_DwellingAge).
\item The fifth data frame is (potentially) empty.
}}
}
\value{
A data frame containing heat loss areas, U-values, thermal
conductivities, area-weighted thermal transmittance (AxU), and conductance
(AxK) for various opaque building elements (\code{BaseFloor}, \code{GrndFloor},
\code{ExpdFloor}, \code{BaseWall}, \code{ExtrWall}, \code{SemiWall}, \code{Roof}, \code{RoomInRoof}
columns) along with construction type and data source.
}
\description{
This function calculates heat loss areas, U-values,
thermal conductivities, and area-weighted thermal transmittance
and conductance for various opaque building elements
(basement floor, ground floor, exposed floor, basement wall,
external wall, semi-exposed wall, roof, and room-in-roof) based on
construction details and reference data.
}
