% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_hlp_openings}
\alias{obtain_hlp_openings}
\title{Obtain U-values and areas for building openings}
\usage{
obtain_hlp_openings(idc, l_inp)
}
\arguments{
\item{idc}{A character string specifying the archetype code.}

\item{l_inp}{A list containing three data frames:
\itemize{
\item The first data frame contains information about the building geometry
(D057_DoorArea, D067_Windows1Area, D069_Windows2Area, D049_RoofWindowArea).
\item The second data frame contains information about the opening types
(D068_Windows1Type, D070_Windows2Type, D058_DoorUValue,
D071_Windows1FrameType, D072_Windows2FrameType).
\item The third data frame contains information about the glazing ratio
(V564_DoubleGlazingRatio) and number of doors (V565_Doors).
}}
}
\value{
A data frame containing U-values and net areas for doors,
windows 1 and 2, and roof windows (\code{Door}, \code{WinA}, \code{WinB}, \code{WinRoof}
columns) along with opening type and frame type.
}
\description{
This function calculates the U-values and net areas for
various building openings (windows, doors, and roof windows) based on
construction details and reference data.
}
