% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_household_bands}
\alias{obtain_household_bands}
\title{Classify household type into bands}
\usage{
obtain_household_bands(household_type)
}
\arguments{
\item{household_type}{A character vector containing household type
information.}
}
\value{
A character vector of the same length as \code{household_type} with
household type bands.
}
\description{
Classifies household type information (\code{household_type}) into
bands.
}
