% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{obtain_idf_output_energy}
\alias{obtain_idf_output_energy}
\title{Obtain IDF Output for Energy}
\usage{
obtain_idf_output_energy(key, output_definitions = l_idf_outputs)
}
\arguments{
\item{key}{The input key.}

\item{output_definitions}{The output definitions. Defaults to \code{l_idf_outputs}.}
}
\value{
A character vector of IDF output variables.
}
\description{
Determines the appropriate IDF output based on the provided key.
}
