% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{obtain_idf_output_lzc}
\alias{obtain_idf_output_lzc}
\title{Obtain IDF Output for Load Zones}
\usage{
obtain_idf_output_lzc(
  load_zone_characteristics,
  zone_information,
  output_definitions_lzc
)
}
\arguments{
\item{load_zone_characteristics}{Load zone characteristics.}

\item{zone_information}{Zone information.}

\item{output_definitions_lzc}{Output definitions for load zones.}
}
\value{
A list containing output variables and meters.
}
\description{
Extracts output variables and meters for load zones based on the presence of solar PV and wind turbines.
}
