% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_income_bands}
\alias{obtain_income_bands}
\title{Classify income into bands}
\usage{
obtain_income_bands(income)
}
\arguments{
\item{income}{A numeric vector containing income values.}
}
\value{
A character vector of the same length as \code{income} with income
bands.
}
\description{
Classifies income values (\code{income}) into bands: "low", "mid",
or "high".
}
