% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_indices}
\alias{obtain_indices}
\title{Prepare TUS data for analysis}
\usage{
obtain_indices(d_tus = s_tus$tbl_individual, path = s_tus$paths$proj)
}
\arguments{
\item{d_tus}{A data frame containing TUS data (default:
s_tus$tbl_individual).}
}
\value{
A data frame containing prepared TUS data.
}
\description{
This function selects relevant variables from a TUS data frame
(\code{d_tus}), re-factors some variables for consistency, and removes
partially empty rows.
}
