% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_indoor_performances}
\alias{obtain_indoor_performances}
\title{Extracts Indoor Performance Data}
\usage{
obtain_indoor_performances(df, len_res = 8760)
}
\arguments{
\item{df}{A data frame containing building performance data.}

\item{len_res}{The length of the simulation results.}
}
\value{
A list containing two data frames:
\itemize{
\item \code{addition}: Heat gain components.
\item \code{removal}: Heat loss components.
}
}
\description{
Extracts annual building sensible heat gain and loss components from a data frame.
}
