% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_inside_heat_transfer_rate}
\alias{obtain_inside_heat_transfer_rate}
\title{Extracts inside heat transfer rates from data}
\usage{
obtain_inside_heat_transfer_rate(d)
}
\arguments{
\item{d}{A data frame containing heat transfer rate data.}
}
\value{
A data frame with information on inside heat transfer rates.
}
\description{
This function extracts inside heat transfer rates for glazing,
opaque surfaces, infiltration, and collates them into a data frame.
Window Transmitted Solar is subtracted so that Glazing heat gain
represents the transfer of all heat through the window excluding
beam and diffuse short-wave solar heat effects. This makes it more
comparable with the opaque surface conduction results.
}
