% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{obtain_insulation_material}
\alias{obtain_insulation_material}
\title{Obtains properties for a specified insulation material}
\usage{
obtain_insulation_material(varThk, lbl.mat = "LF_i_fibre_roll")
}
\arguments{
\item{varThk}{A numeric value representing the insulation thickness in millimeters.}

\item{lbl.mat}{A character value specifying the desired material name (default: "LF_i_fibre_roll").}
}
\value{
A list containing:
\itemize{
\item table: A data frame containing material properties.
\item idf: A character vector representing the IDF content for the material.
}
}
\description{
Extracts material properties and generates IDF content for a given insulation type.
}
