% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-loss.R
\name{obtain_list_zone_summary}
\alias{obtain_list_zone_summary}
\title{Obtain a list of zone summaries}
\usage{
obtain_list_zone_summary(archetype_info, reference_data, zone_data)
}
\arguments{
\item{archetype_info}{A data frame containing archetype
information (vFlr, vBsm, vRff, vFlt).}

\item{reference_data}{A reference data frame used by obtain_archetype_level.}

\item{zone_data}{A data frame containing zone
information (zone, room_function, area__m2, volume__m3, glazed).}
}
\value{
A list containing:
\itemize{
\item all_zones: All zones in the data frame.
\item gather: Zones classified as living or dining rooms.
\item rest: Zones classified as bedrooms.
\item cooking: Zones classified as kitchens.
\item hygiene: Zones classified as bathrooms or water-related spaces.
\item other: Zones not classified in the previous categories.
\item size: Total number of zones.
}
}
\description{
This function processes information about an archetype
and its zones, creating a list containing various zone-related data.
}
