% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_material_summary}
\alias{obtain_material_summary}
\title{Extracts Material Summary Information}
\usage{
obtain_material_summary(df)
}
\arguments{
\item{df}{A data frame containing building construction and material data.}
}
\value{
A list containing four data frames:
\itemize{
\item \code{construction}: Construction information.
\item \code{materials}: Material information.
\item \code{glazing}: Glazing information.
\item \code{opaque}: Opaque envelope information.
}
}
\description{
Extracts construction, material, glazing, and opaque envelope information from a data frame.
}
