% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_matrix_size}
\alias{obtain_matrix_size}
\title{Obtain dimensions of matrix}
\usage{
obtain_matrix_size(id, l)
}
\arguments{
\item{id}{selected layout}

\item{l}{list of available layouts}
}
\value{
list of parsed attributes
}
\description{
Obtain dimensions of matrix
}
