% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{obtain_most_frequent}
\alias{obtain_most_frequent}
\title{Obtains the most frequent row in a data frame}
\usage{
obtain_most_frequent(d)
}
\arguments{
\item{d}{Data frame to process.}
}
\value{
The most frequent row without the \code{V002_DwellingWeight} column.
}
\description{
This function selects the row with the highest value in the \code{V002_DwellingWeight} column
and removes that column from the result.
}
