% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_occupants_presence}
\alias{obtain_occupants_presence}
\title{Obtains occupant presence data}
\usage{
obtain_occupants_presence(
  df_cmp,
  df_sum,
  df_inf,
  df_ind,
  df_dry,
  df_occ,
  k_sel,
  k_met,
  path_proj = the$.enhub.paths
)
}
\arguments{
\item{df_cmp}{A data frame containing comprehensive household data.}

\item{df_sum}{A data frame containing summary household data.}

\item{df_inf}{A data frame containing individual information.}

\item{df_ind}{A data frame containing individual data.}

\item{df_dry}{A data frame containing diary data.}

\item{df_occ}{A data frame containing occupant information.}

\item{k_sel}{A character vector specifying the selection criteria.}

\item{k_met}{A character vector specifying the method.}

\item{path_proj}{The path to the project directory (default:
the$.enhub.paths).}
}
\value{
A data frame containing presence data for all occupants.
}
\description{
Extracts household characteristics, matches a profile,
generates presence matrices, and merges the resulting presence data for
all occupants.
}
