% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_region_bands}
\alias{obtain_region_bands}
\title{Classify region into bands}
\usage{
obtain_region_bands(region)
}
\arguments{
\item{region}{A character vector containing region information.}
}
\value{
A character vector of the same length as \code{region} with region
bands.
}
\description{
Classifies region information (\code{region}) into standard UK
region bands.
}
