% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_site_air_profile}
\alias{obtain_site_air_profile}
\title{Extracts Site Air Temperature Profile}
\usage{
obtain_site_air_profile(d)
}
\arguments{
\item{d}{A data frame containing weather data with columns \code{Date.Time} and
\code{Outdoor.Air.Drybulb.Temperature..C..Hourly.}.}
}
\value{
A JSON string representing the extracted temperature data.
}
\description{
This function extracts outdoor dry bulb temperature from a given dataset,
renames columns, and converts the data to JSON format.
}
