% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_surface_summary}
\alias{obtain_surface_summary}
\title{Extracts surface information from data frame}
\usage{
obtain_surface_summary(df)
}
\arguments{
\item{df}{A data frame containing surface information.}
}
\value{
A list containing a data frame with surface characteristics.
}
\description{
This function extracts surface information from a data frame,
cleans and reformats the data, and returns a summary.
}
