% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{obtain_surfaces_info}
\alias{obtain_surfaces_info}
\title{Obtain generated list of surface parameters}
\usage{
obtain_surfaces_info(l_bl, l_gl, l_dr)
}
\arguments{
\item{l_bl}{list of opaque surfaces}

\item{l_gl}{list of glazing surfaces}

\item{l_dr}{list of door surfaces}
}
\value{
a unified list of surface parameters
}
\description{
Obtain generated list of surface parameters
}
\examples{
\dontrun{
obtain_surfaces_info(l_blocks, l_glazing, l_doors)
}
}
