% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_tenure_bands}
\alias{obtain_tenure_bands}
\title{Classify tenure type into bands}
\usage{
obtain_tenure_bands(tenure_type)
}
\arguments{
\item{tenure_type}{A character vector containing tenure type information.}
}
\value{
A character vector of the same length as \code{tenure_type} with tenure
bands.
}
\description{
Classifies tenure type (\code{tenure_type}) into bands: "Private",
"Social", or NA.
}
