% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{obtain_typology_bands}
\alias{obtain_typology_bands}
\title{Classify dwelling type into bands}
\usage{
obtain_typology_bands(dwelling_type)
}
\arguments{
\item{dwelling_type}{A character vector containing dwelling type
information.}
}
\value{
A character vector of the same length as \code{dwelling_type} with
dwelling type bands.
}
\description{
Classifies dwelling type (\code{dwellingType}) into bands: "House"
or "Apartment".
}
