% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_water_services}
\alias{obtain_water_services}
\title{Extracts Water Services Information}
\usage{
obtain_water_services(df, dd)
}
\arguments{
\item{df}{A data frame containing water heating energy data.}

\item{dd}{A data frame containing water volume data.}
}
\value{
A list containing:
\itemize{
\item \code{dhw_services}: A data frame with water heating services and their contribution.
\item \code{dwh_demand__kWh}: Total water heating energy demand in kWh.
\item \code{water_volume}: Total water volume in m3.
}
}
\description{
Extracts water heating energy demand and water volume from data frames.
}
