% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_zone_summary}
\alias{obtain_zone_summary}
\title{Extracts Zone Summary Information}
\usage{
obtain_zone_summary(df)
}
\arguments{
\item{df}{A data frame containing zone data.}
}
\value{
A list containing two data frames:
\itemize{
\item \code{summary}: Zone summary information.
\item \code{key}: Zone key information.
}
}
\description{
Extracts zone summary and information data from a given data frame.
}
