% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{obtain_zone_temperature_ddy}
\alias{obtain_zone_temperature_ddy}
\title{Extracts Zone Temperature Data by Day/Year}
\usage{
obtain_zone_temperature_ddy(dp)
}
\arguments{
\item{dp}{A data frame containing zone temperature data.}
}
\value{
A list of data frames, each containing zone temperature data for a specific day/year item.
}
\description{
Extracts zone temperature data for each day/year item.
}
