% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{parse_header}
\alias{parse_header}
\title{Parse Header Elements from EP+ Summary File}
\usage{
parse_header(txth)
}
\arguments{
\item{txth}{A character string representing the header of an EP+ summary file.}
}
\value{
A character vector of parsed header elements.
}
\description{
This function takes a header string from an EP+ summary file and
parses it into a character vector of header elements.

The function performs the following transformations on the input string:
\enumerate{
\item Replaces leading ",," with "ctrl,item,".
\item Replaces "#" with "no_".
\item Replaces "\{" and "\/" with "__".
\item Removes "\}".
\item Replaces "-" and whitespace with "_".
\item Converts the string to lowercase.
\item Splits the string into a character vector using "," as the delimiter.
}
}
