% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{parse_space_heating_group}
\alias{parse_space_heating_group}
\title{Parse Space Heating Group}
\usage{
parse_space_heating_group(energy_system_code)
}
\arguments{
\item{energy_system_code}{The energy system code.}
}
\value{
The space heating group.
}
\description{
Determines the space heating group based on the provided energy system code.
}
