% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-header.R
\name{populate_idf_header}
\alias{populate_idf_header}
\title{Populates the IDF Header with Building Information}
\usage{
populate_idf_header(
  id,
  l_sum,
  l_ids = the$core_index,
  ehs_data = s_ehs_2011_ext,
  ons_data = s_ons,
  tus_data = s_tus,
  chm_data = s_chm
)
}
\arguments{
\item{id}{The ID of the building.}

\item{l_sum}{Summary data for the building.}

\item{l_ids}{A list of building IDs (default is \code{the$core_index}).}

\item{ehs_data}{A list of EHS data sets (default is \code{s_ehs_2011_ext}).}

\item{ons_data}{A list of ONS data sets (default is \code{s_ons}).}

\item{tus_data}{A list of TUS data sets (default is \code{s_tus}).}

\item{chm_data}{A list of CHM reference data (default is \code{s_chm}).}
}
\value{
A list containing the card and IDF data.
}
\description{
This function populates the IDF header with
detailed building information, including characteristics,
energy systems, and construction details. It merges data
from various sources and updates the building characteristics
based on predefined defaults.
}
