% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-sections.R
\name{populate_idf_section_E}
\alias{populate_idf_section_E}
\title{Populates the IDF section E with data based on provided parameters.}
\usage{
populate_idf_section_E(
  hid,
  d_ids = the$core_index,
  l_ehs = s_ehs_2011_ext,
  l_rom = s_ehs_2011_rom,
  l_sup = l_idf_blocks
)
}
\arguments{
\item{hid}{Character string representing the housing code.}

\item{d_ids}{Dataframe containing housing code information.}

\item{l_ehs}{List of EHS data.}

\item{l_rom}{Unused argument.}

\item{l_sup}{List of IDF blocks.}
}
\value{
A list containing the populated IDF section E.
}
\description{
This function processes housing data, calculates dwelling geometry, and defines
thermal zone groups to populate the IDF section E.
}
