% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf-sections.R
\name{populate_idf_section_F}
\alias{populate_idf_section_F}
\title{Populates the IDF section F with data based on provided parameters.}
\usage{
populate_idf_section_F(
  hid,
  l_weather,
  l_sched,
  l_dims,
  d_ids = the$core_index,
  l_ehs = s_ehs_2011_ext
)
}
\arguments{
\item{hid}{Character string representing the housing code.}

\item{l_weather}{List containing weather data.}

\item{l_sched}{List containing schedule data.}

\item{l_dims}{List containing building dimensions data.}

\item{d_ids}{Dataframe containing housing code information (default: the$core_index).}

\item{l_ehs}{List containing EHS data (default: s_ehs_2011_ext).}
}
\value{
A list containing a summary and IDF sections.
}
\description{
This function processes housing data, calculates building characteristics,
and generates IDF sections for internal loads, infiltration, mixing, and
a summary of key parameters.
}
