% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{predict_room_info}
\alias{predict_room_info}
\title{Predicts room information based on input parameters.}
\usage{
predict_room_info(
  l_models,
  i_type,
  i_area,
  i_nrooms = NULL,
  i_nstoreys = NULL,
  d_ref = s_ehs_2011_rom
)
}
\arguments{
\item{l_models}{A list of prediction models.}

\item{i_type}{Character string representing the dwelling type.}

\item{i_area}{Numeric value representing the floor area.}

\item{i_nrooms}{Optional numeric value specifying the number of rooms.}

\item{i_nstoreys}{Optional numeric value specifying the number of storeys.}

\item{d_ref}{Data frame containing reference data for room information.}
}
\value{
A JSON string containing predicted room information.
}
\description{
Uses pre-trained models to predict room-related information for a given dwelling.
}
