% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{prepare_enhub_output_dir}
\alias{prepare_enhub_output_dir}
\title{Prepares Output Directories}
\usage{
prepare_enhub_output_dir()
}
\description{
Creates a set of output directories for EnHub simulations.
}
\details{
This function creates the necessary output directories for
EnHub simulations, including directories for simulations,
reports, sensitivity analysis, scenarios, and selection.
}
