% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{process_output_resolution}
\alias{process_output_resolution}
\title{Processes output resolution}
\usage{
process_output_resolution(df)
}
\arguments{
\item{df}{A data frame containing output data with a column named "Date.Time".}
}
\value{
A list containing the following elements:
\itemize{
\item \code{resolution}: A character string indicating the data resolution (hourly, daily, monthly, or timestep).
\item \code{ddys}: A numeric vector indicating the indices of days with incomplete data.
\item \code{days}: A numeric vector indicating the indices of full days.
}
}
\description{
This function determines the data resolution based on the column name,
processes the "Date.Time" column, and creates relevant date-related vectors.
}
