% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_mkdir}
\alias{r_mkdir}
\title{Creates a Directory}
\usage{
r_mkdir(pathNew, myOS = r_uname())
}
\arguments{
\item{pathNew}{The path to the directory to be created.}

\item{myOS}{The operating system (defaults to the result of \code{r_uname()}).}
}
\description{
Creates a directory at the specified path, handling
platform differences.
}
\details{
The function creates a directory at the specified \code{pathNew}. It handles
Windows and non-Windows systems differently. For Windows, it normalizes
the path and uses \code{dir.create} to create the directory. For other
systems, it uses the \code{mkdir -p} command to create the directory and
any necessary parent directories.
}
