% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_mv}
\alias{r_mv}
\title{Moves a File or Directory}
\usage{
r_mv(varFrom, varTo, myOS = r_uname())
}
\arguments{
\item{varFrom}{The path to the file or directory to be moved.}

\item{varTo}{The destination path for the file or directory.}

\item{myOS}{The operating system (defaults to the result of \code{r_uname()}).}
}
\description{
Moves a file or directory from one location to another,
handling platform differences.
}
\details{
The function moves a file or directory from \code{varFrom} to \code{varTo}. It
handles Windows and non-Windows systems differently. For Windows, it
uses the \code{MOVE /Y} command to move the file or directory. For other
systems, it uses the \code{mv} command to move the file or directory.
}
