% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_rm}
\alias{r_rm}
\title{Removes a Directory}
\usage{
r_rm(pathNew, myOS = r_uname())
}
\arguments{
\item{pathNew}{The path to the directory to be removed.}

\item{myOS}{The operating system (defaults to the result of \code{r_uname()}).}
}
\description{
Removes a directory and its contents at the specified
path, handling platform differences.
}
\details{
The function removes a directory and its contents at the specified
\code{pathNew}. It handles Windows and non-Windows systems differently.
For Windows, it normalizes the path and uses \code{unlink} to remove the
directory and its contents. For other systems, it uses the \code{rm -rf}
command to remove the directory and its contents.
}
