% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_uname}
\alias{r_uname}
\title{Determines the operating system}
\usage{
r_uname()
}
\value{
A character string representing the operating system ("win", "mac", "linux", or "hpc").
}
\description{
This function identifies the user's operating system.
}
\details{
The function uses a combination of checks based on \code{Sys.info()} and \code{.Platform$OS.type} to identify the operating system.
It returns a specific string based on the detected OS:
\itemize{
\item "win" for Windows
\item "mac" for macOS
\item "linux" for Linux (assuming GNU)
\item "hpc" for Linux (other distributions)
}

If the function cannot identify the OS, it throws an error message ("Mystery Machine").
}
\seealso{
Other R-environment helpers: 
\code{\link{r_whoami}()}
}
\concept{R-environment helpers}
