% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_unzip}
\alias{r_unzip}
\title{Unzips a Compressed File}
\usage{
r_unzip(pathNew, myOS = r_uname(), ...)
}
\arguments{
\item{pathNew}{The path to the compressed file (without the .gz extension).}

\item{myOS}{The operating system (defaults to the result of \code{r_uname()}).}

\item{...}{Additional arguments to be passed to the unzip function.}
}
\description{
Unzips a compressed file, handling platform differences.
}
\details{
The function unzips the specified compressed file (with a .gz extension)
using either the built-in \code{gunzip} function (for non-Windows systems) or
a custom unzip function (for Windows) based on the specified \code{pathZip}
in the \code{the$.enhub.config} settings.
}
