% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_whoami}
\alias{r_whoami}
\title{Gets User Information}
\usage{
r_whoami()
}
\value{
A character string containing a welcome message with the user name.
}
\description{
Retrieves the user name for the current operating system.
}
\details{
The function determines the operating system using \code{r_uname}.
If the OS is not Windows, it uses the \code{whoami} command to get the user name.
For Windows, it defaults to "windows user".
}
\seealso{
Other R-environment helpers: 
\code{\link{r_uname}()}
}
\concept{R-environment helpers}
