% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{r_zip}
\alias{r_zip}
\title{Zips a Directory}
\usage{
r_zip(path_to_zip)
}
\arguments{
\item{path_to_zip}{The path to the file/directory to be zipped.}
}
\description{
Zips a directory, handling platform differences.
}
\details{
The function zips the specified directory using either the
built-in \code{gzip} function (for non-Windows systems) or
a custom zip function (for Windows) based on the specified \code{pathZip}
in the \code{the$.enhub.config} settings.
}
