% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{randomise_profiles}
\alias{randomise_profiles}
\title{Randomises profiles used for scheduling}
\usage{
randomise_profiles(k = "det")
}
\arguments{
\item{k}{Character string indicating the type of randomisation.
Default is "det" for deterministic behaviour.
If "sto", sets the seed to NULL for stochastic behavior.}
}
\description{
Sets the random seed for profile randomisation.
}
