% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{read_idf}
\alias{read_idf}
\title{Reads an IDF File}
\usage{
read_idf(dir_path, file_path)
}
\arguments{
\item{dir_path}{The directory path of the IDF file.}

\item{file_path}{The file name of the IDF file.}
}
\value{
A character vector containing the lines of the IDF file.
}
\description{
Reads an IDF file and returns its contents as a character vector.
}
\details{
This function reads an IDF file from the specified directory and file path.
The file content is read as a character vector, with tabs removed.
}
