% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{read_idf_outcomes}
\alias{read_idf_outcomes}
\title{Reads general properties from an IDF file}
\usage{
read_idf_outcomes(idf, path_ref = the$.enhub.paths$eval$B)
}
\arguments{
\item{idf}{Character string specifying the IDF file name.}

\item{path_ref}{Character string specifying the path to the stored batch.
Defaults to \code{the$.enhub.paths$eval$B}.}
}
\value{
A character vector containing the lines of the extracted data.
}
\description{
Extracts general properties parsed to an IDF file from a specified path.
The function cleans the read data by removing tabs and certain characters.
}
\seealso{
Other EnergyPlus evaluation: 
\code{\link{convert_idf_to_json}()},
\code{\link{extract_simulation_outputs}()},
\code{\link{generate_idf}()},
\code{\link{generate_idf_multiple}()},
\code{\link{read_input_file}()},
\code{\link{read_input_summary}()},
\code{\link{run_energy_plus}()},
\code{\link{simulate_available_idfs}()}
}
\concept{EnergyPlus evaluation}
