% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{read_input_summary}
\alias{read_input_summary}
\title{Reads input summary JSON file from a stored batch}
\usage{
read_input_summary(idf, path_ref = the$.enhub.paths$eval$A)
}
\arguments{
\item{idf}{Character string specifying the IDF file name.}

\item{path_ref}{Character string specifying the path to the stored batch.
Defaults to \code{the$.enhub.paths$eval$A}.}
}
\value{
An R object representing the parsed JSON content.
}
\description{
This function reads a JSON input summary file from a specified path
within a stored batch. The function parses the JSON content and returns
it as an R object.
}
\seealso{
Other EnergyPlus evaluation: 
\code{\link{convert_idf_to_json}()},
\code{\link{extract_simulation_outputs}()},
\code{\link{generate_idf}()},
\code{\link{generate_idf_multiple}()},
\code{\link{read_idf_outcomes}()},
\code{\link{read_input_file}()},
\code{\link{run_energy_plus}()},
\code{\link{simulate_available_idfs}()}
}
\concept{EnergyPlus evaluation}
