% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{reassign_glazing}
\alias{reassign_glazing}
\title{Reassigns glazing types}
\usage{
reassign_glazing(li, dhl = s_ehs_2011_ext$other.heat.loss.elements)
}
\arguments{
\item{li}{A character string indicating the level of glazing: "ALL", "MORE_THAN_HALF", "LESS_THAN_HALF", or any other value.}

\item{dhl}{A data frame containing heat loss elements (default: s_ehs_2011_ext$other.heat.loss.elements).}
}
\value{
The modified data frame with reassigned glazing types.
}
\description{
Reassigns glazing types based on the provided level of glazing.
}
