% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{reassign_thermostat_temperature}
\alias{reassign_thermostat_temperature}
\title{Reassigns thermostat set points based on input category}
\usage{
reassign_thermostat_temperature(li, l_par = the$GLOBAL_PAR)
}
\arguments{
\item{li}{Character string indicating the temperature category.}

\item{l_par}{List of parameters, defaulting to \code{the$GLOBAL_PAR}.}
}
\value{
The modified \code{l_par} list with updated set points.
}
\description{
This function reassigns the \code{set_point_heating} and \code{set_point_cooling}
parameters within the \code{l_par} list based on the provided input category \code{li}.
}
