% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{reassign_wall_type}
\alias{reassign_wall_type}
\title{Reassigns wall type based on input value}
\usage{
reassign_wall_type(li, dhl = s_ehs_2011_ext$geometry)
}
\arguments{
\item{li}{Character string indicating the wall type.}

\item{dhl}{Spatial data frame containing the \code{D077_ExternalWallConstruction} field.
Defaults to \code{s_ehs_2011_ext$geometry}.}
}
\value{
The modified \code{dhl} data frame with updated wall types.
}
\description{
This function reassigns the \code{D077_ExternalWallConstruction} field in the provided
\code{dhl} data frame based on the input value \code{li}. The mapping between \code{li} and
wall types is defined within the function.
}
\references{
s_chm$CHMtbl_EHS_CHM_wallType\link{1:16,}
}
