% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{rename_indexed_simulation}
\alias{rename_indexed_simulation}
\title{Renames a simulation based on index and UUID}
\usage{
rename_indexed_simulation(i, df)
}
\arguments{
\item{i}{Index of the simulation to rename.}

\item{df}{Data frame containing simulation data.}
}
\value{
The modified data frame.
}
\description{
This function renames a simulation in a data frame based on its index and UUID.
}
