% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{rename_model_with_uuid}
\alias{rename_model_with_uuid}
\title{Renames model files with UUID}
\usage{
rename_model_with_uuid(idn)
}
\arguments{
\item{idn}{The original file ID/index}
}
\description{
Renames files with a specified ID in multiple
directories using a UUID.

A UUID is a universally unique identifier, ie. a 128-bit
number used to identify information in computer systems
→ \href{https://www.wikidata.org/entity/Q195284}{wikidata}
}
\details{
This function renames files with a given ID to a
new ID (UUID) in specified directories. It iterates through
the directories in \code{the$.enhub.paths$eval}, constructs the
old and new file paths, and renames the file using the
\code{r_mv} function.
}
\examples{
\dontrun{
# Assuming `id_model` exists and was used to generate the model
rename_with_uuid(id_model)
}

}
