% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{replace_na_with_last}
\alias{replace_na_with_last}
\title{Replace NA Values with Previous Non-Missing Values}
\usage{
replace_na_with_last(x, a = !is.na(x))
}
\arguments{
\item{x}{A numeric vector containing potentially missing values (NA).}

\item{a}{A logical vector of the same length as \code{x}, indicating which elements
are not missing (TRUE) or missing (FALSE). Defaults to checking for missing
values directly in \code{x}.}
}
\value{
A numeric vector with NA values replaced by the last non-missing value.
}
\description{
Replaces NA values in a vector with the last non-missing value encountered.
}
