% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{reset_folders}
\alias{reset_folders}
\title{Resets specified folders}
\usage{
reset_folders(path_names = NULL)
}
\arguments{
\item{path_names}{A character vector specifying the folders to reset.
Possible values include "simulation", "scenarios", "sensitivity", "profiles".}
}
\description{
This function resets a set of specified folders by
removing their contents and recreating them.
}
\details{
The function identifies the folders to be reset based on the
provided \code{path_names} argument. It then removes the contents of these
folders and recreates them. Finally, it prints a list of the
recreated folders.
}
\examples{
\dontrun{
# Reset simulation and scenarios folders only
reset_folders(c("simulation", "scenarios"))

# Reset all folders
reset_folders()
}

}
