% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-beis.R
\docType{data}
\name{s_beis}
\alias{s_beis}
\title{Data from BEIS}
\format{
A list with 4  elements:
\itemize{
\item gases.total: A tibble containing data from GHGtable1, which likely includes total greenhouse gas emissions data.
\item sources.domestic: A tibble containing data from GHGtable2, which likely includes data on domestic sources of greenhouse gas emissions.
\item enduses.domestic: A tibble containing data from GHGtable3, which likely includes data on domestic end uses of greenhouse gas emissions.
\item fuels.total: A tibble containing data from GHGtable4, which likely includes total data on fuels contributing to greenhouse gas emissions.
}
}
\usage{
s_beis
}
\description{
This object contains the data retrieved from the BEIS sources
}
\examples{
\dontrun{
# Displaying the whole list
s_beis
}
}
\keyword{datasets}
