% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bredem.R
\docType{data}
\name{s_bredem}
\alias{s_bredem}
\title{Data from BREDEM}
\format{
A list with 3  elements:
\itemize{
\item FluxConstants (data.frame): A data frame containing \emph{'Table 15`: Constants for calculation of solar flux on vertical and inclined surfaces'}
\item GlobalIrradiance (data.frame): A data frame containing \emph{'Table A1`: Mean global solar irradiance (W/m2) on a horizontal plane, and latitude (° North)'}
\item HorizontalRadiation (data.frame): A data frame containing \emph{'Table D13`: Monthly Average Horizontal Solar Radiation (W/m2)'}
}
}
\usage{
s_bredem
}
\description{
This object contains the data retrieved from the BREDEM sources
}
\examples{
\dontrun{
# Displaying the whole list
s_bredem
}
}
\seealso{
Other core data sources: 
\code{\link{s_chm}},
\code{\link{s_chm_reference}},
\code{\link{s_ehs_2011_ext}},
\code{\link{s_ehs_2011_hhd}},
\code{\link{s_ehs_2011_rom}},
\code{\link{s_sap}},
\code{\link{s_tus}}
}
\concept{core data sources}
\keyword{datasets}
