% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chm.R
\docType{data}
\name{s_chm_reference}
\alias{s_chm_reference}
\title{CHM Reference Data}
\format{
A list of 21 data frames:
\describe{
\item{_ConsList}{Data frame with 5 observations and 16 variables: V1, V2, V3, V4, V5, V6, V7, V8, V9, V10, V11, V12, V13, V14, V15, V16.}
\item{_MatList}{Data frame with 21 observations and 6 variables: Name, Roughness, Thickness..m., Conductivity..W.mK., Density..kg.m3., SpecificHeat..J.kgK..}
\item{BasementGroundFloorInsulation}{Data frame with 11 observations and 3 variables: EngWal, Scot, Nire.}
\item{ExposedFloorConstruction}{Data frame with 3 observations and 2 variables: Exposed.Floor.Construction, Exposed.Floor.K.value..kJ.m2K..}
\item{ExposedFloorUvalue}{Data frame with 12 observations and 3 variables: AgeBand, X1.Exposed.Floor.U.value..W.m2K., X2.Exposed.Floor.U.value..W.m2K..}
\item{FloorUParameters}{Data frame with 10 observations and 3 variables: name, value, units.}
\item{GroundFloorConstruction}{Data frame with 4 observations and 2 variables: Type, Basement.Ground.Floor.K.Value..kJ.m2K..}
\item{InternalConstructionK}{Data frame with 4 observations and 2 variables: Internal.Wall.Construction.and.K.value..kJ.m2K....SAP.Table.1e, K.value.}
\item{PartyCeiling}{Data frame with 2 observations and 2 variables: name, value.}
\item{PartyFloor}{Data frame with 7 observations and 2 variables: name, value.}
\item{PartyWallConstruction}{Data frame with 7 observations and 3 variables: Party.Wall.Construction.U.value..W.m2K..and.K.value..kJ.m2K....SAP.Table.3.6.and.Table.1e, K.value, U.value.}
\item{Roof}{Data frame with 12 observations and 1 variable: RoofValue.}
\item{RoofConstruction}{Data frame with 4 observations and 2 variables: name, value.}
\item{RoofRoom}{Data frame with 3 observations and 1 variable: name.}
\item{RoofU}{Data frame with 11 observations and 7 variables: UvKn.Pitched, UvKn.Thatched, UvUn.Pitched, UvUn.Pitched.1, UvUn.Flat, UvAt.Pitched, UvAt.Thatched.}
\item{WallType}{Data frame with 16 observations and 59 variables: id, Wall.Construction.Type, Wall.K.value..kJ.m2K....calculated.based.on.assumptions.in.construction.details, X.p...density..c...specific.heat.capacity., wllThick_m_1, wllThick_m_2, wllThick_m_3, wllThick_m_4, wllThick_m_5, wllThick_m_6, wllThick_m_7, wllThick_m_8, wllThick_m_9, wllThick_m_10, wllThick_m_11, wllUv_Wm2K_EngWal_01, wllUv_Wm2K_EngWal_02, wllUv_Wm2K_EngWal_03, wllUv_Wm2K_EngWal_04, wllUv_Wm2K_EngWal_05, wllUv_Wm2K_EngWal_06, wllUv_Wm2K_EngWal_07, wllUv_Wm2K_EngWal_08, wllUv_Wm2K_EngWal_09, wllUv_Wm2K_EngWal_10, wllUv_Wm2K_EngWal_11, wllUv_Wm2K_Sco_01, wllUv_Wm2K_Sco_02, wllUv_Wm2K_Sco_03, wllUv_Wm2K_Sco_04, wllUv_Wm2K_Sco_05, wllUv_Wm2K_Sco_06, wllUv_Wm2K_Sco_07, wllUv_Wm2K_Sco_08, wllUv_Wm2K_Sco_09, wllUv_Wm2K_Sco_10, wllUv_Wm2K_Sco_11, wllUvSem_Wm2K_EngWal_01, wllUvSem_Wm2K_EngWal_02, wllUvSem_Wm2K_EngWal_03, wllUvSem_Wm2K_EngWal_04, wllUvSem_Wm2K_EngWal_05, wllUvSem_Wm2K_EngWal_06, wllUvSem_Wm2K_EngWal_07, wllUvSem_Wm2K_EngWal_08, wllUvSem_Wm2K_EngWal_09, wllUvSem_Wm2K_EngWal_10, wllUvSem_Wm2K_EngWal_11, wllUvSem_Wm2K_Sco_01, wllUvSem_Wm2K_Sco_02, wllUvSem_Wm2K_Sco_03, wllUvSem_Wm2K_Sco_04, wllUvSem_Wm2K_Sco_05, wllUvSem_Wm2K_Sco_06, wllUvSem_Wm2K_Sco_07, wllUvSem_Wm2K_Sco_08, wllUvSem_Wm2K_Sco_09, wllUvSem_Wm2K_Sco_10, wllUvSem_Wm2K_Sco_11.}
\item{Window}{Data frame with 9 observations and 14 variables: Window.Type, Type, Uv_Wood_NoC, Uv_Metal_NoC, Uv_uPVC_NoC, Uv_Wood_Eff, Uv_Metal_Eff, Uv_uPVC_Eff, Roof_Uv_Wood_NoC, Roof_Uv_Metal_NoC, Roof_Uv_uPVC_NoC, Roof_Uv_Wood_Eff, Roof_Uv_Metal_Eff, Roof_Uv_uPVC_Eff.}
\item{WindowFrameType}{Data frame with 4 observations and 1 variable: name.}
\item{WindowOrientation}{Data frame with 7 observations and 3 variables: Ori, grads, rads.}
\item{WindowOvershading}{Data frame with 5 observations and 1 variable: Window.Overshading.}
\item{WindowTransmittance}{Data frame with 9 observations and 3 variables: Window.Transmittance, Solar, Light.}
}
}
\usage{
s_chm_reference
}
\description{
This list contains various data frames related to construction materials, insulation, floor construction, wall types, windows, and other building components.
}
\examples{
\dontrun{
# Accessing the _ConsList data frame
s_chm_reference$`_ConsList`

# Summary of the _MatList data frame
summary(s_chm_reference$`_MatList`)
}
}
\seealso{
Other core data sources: 
\code{\link{s_bredem}},
\code{\link{s_chm}},
\code{\link{s_ehs_2011_ext}},
\code{\link{s_ehs_2011_hhd}},
\code{\link{s_ehs_2011_rom}},
\code{\link{s_sap}},
\code{\link{s_tus}}
}
\concept{core data sources}
\keyword{datasets}
