% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dukes.R
\docType{data}
\name{s_dukes}
\alias{s_dukes}
\title{Data from DUKES}
\format{
A list containing various datasets related to energy supply and usage.
\itemize{
\item supply: A tibble containing data on energy supply from 1998 to 2018.
\item enduse: A tibble containing data on energy end use from 1970 to 2018.
\item enduse_all: A tibble containing comprehensive data on energy end use from 1970 to 2018.
\item electricity: A tibble containing data on electricity usage from 1970 to 2018.
}
}
\usage{
s_dukes
}
\description{
This object contains the data retrieved from the DUKES sources
}
\examples{
\dontrun{
# Displaying the whole list
s_dukes
}
}
\keyword{datasets}
