% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ecuk.R
\docType{data}
\name{s_ecuk_2012}
\alias{s_ecuk_2012}
\title{Data from ECUK 2012}
\format{
list containing various datasets related to energy consumption and usage in the UK for the year 2012.
\itemize{
\item population_x1000.der: A tibble containing data on population in thousands.
\item cold_x1000.der: A tibble containing data on cold-related metrics in thousands.
\item centralheat_x1000.der: A tibble containing data on central heating in thousands.
\item boiler_x1000.der: A tibble containing data on boilers in thousands.
\item EnergyUse.der: A tibble containing data on energy use.
\item intensity.der: A tibble containing data on energy intensity.
\item intensityNormal.melt: A tibble containing normalized data on energy intensity.
}
}
\usage{
s_ecuk_2012
}
\description{
This object contains the data retrieved from the ECUK sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ecuk_2012
}
}
\keyword{datasets}
