% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ecuk.R
\docType{data}
\name{s_ecuk_2014}
\alias{s_ecuk_2014}
\title{Data from ECUK 2014}
\format{
A list containing various datasets related to energy consumption and usage in the UK for the year 2014.
\itemize{
\item fuels_mtoe__derived: A tibble containing data on fuels in million tonnes of oil equivalent (Mtoe), derived.
\item fuels_ktoe.der: A tibble containing data on fuels in thousand tonnes of oil equivalent (ktoe), derived.
\item systems_mtoe__derived: A tibble containing data on energy systems in million tonnes of oil equivalent (Mtoe), derived.
\item temperature_c.der: A tibble containing data on temperature in degrees Celsius, derived.
\item electricityGas.der: A tibble containing data on electricity and gas usage, derived.
\item appliances_ktoe.der: A tibble containing data on appliances in thousand tonnes of oil equivalent (ktoe), derived.
\item hhd.appl.groups: A tibble containing data on household appliance groups.
\item averageDay.der: A tibble containing data on average daily energy usage, derived.
\item sectorContribution.der: A tibble containing data on sector contributions to energy usage, derived.
\item domesticFuel.der: A tibble containing data on domestic fuel usage, derived.
}
}
\usage{
s_ecuk_2014
}
\description{
This object contains the data retrieved from the ECUK sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ecuk_2014
}
}
\keyword{datasets}
