% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ehs.R
\docType{data}
\name{s_ehs_2011_ext}
\alias{s_ehs_2011_ext}
\title{Data from EHS}
\format{
A list containing various datasets related to the English Housing Survey (EHS) for the year 2011:
\itemize{
\item dwelling.and.household.information: Data on dwelling and household information.
\item geometry: Data on the geometry of the dwellings.
\item ventilation: Data on ventilation systems and parameters.
\item other.heat.loss.elements: Data on other elements contributing to heat loss.
\item space.heating: Data on space heating systems and usage.
\item hot.water.system: Data on hot water systems.
\item low.energy.lighting: Data on low energy lighting usage.
\item complementary: Complementary data related to the survey.
\item summarised: Summarised data from the survey.
\item common: Common data elements used across different datasets.
}
}
\usage{
s_ehs_2011_ext
}
\description{
This object contains the data retrieved from the English Housing Survey sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ehs_2011_ext
}
}
\seealso{
Other core data sources: 
\code{\link{s_bredem}},
\code{\link{s_chm}},
\code{\link{s_chm_reference}},
\code{\link{s_ehs_2011_hhd}},
\code{\link{s_ehs_2011_rom}},
\code{\link{s_sap}},
\code{\link{s_tus}}
}
\concept{core data sources}
\keyword{datasets}
