% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ehs.R
\docType{data}
\name{s_ehs_2011_hhd}
\alias{s_ehs_2011_hhd}
\title{Data from EHS}
\format{
A list containing various datasets related to the English Housing Survey (EHS) for the year 2011, focusing on household data:
\itemize{
\item aacode: Data on a specific aspect of household information by aacode
}
}
\usage{
s_ehs_2011_hhd
}
\description{
This object contains the data retrieved from the English Housing Survey sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ehs_2011_hhd
}
}
\seealso{
Other core data sources: 
\code{\link{s_bredem}},
\code{\link{s_chm}},
\code{\link{s_chm_reference}},
\code{\link{s_ehs_2011_ext}},
\code{\link{s_ehs_2011_rom}},
\code{\link{s_sap}},
\code{\link{s_tus}}
}
\concept{core data sources}
\keyword{datasets}
