% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ehs.R
\docType{data}
\name{s_ehs_2011_rom}
\alias{s_ehs_2011_rom}
\title{Data from EHS}
\format{
A list containing various datasets related to room characteristics from the English Housing Survey (EHS) for the year 2011.
\itemize{
\item aacode: Unique identifier for the dwelling.
\item dwtype7x: Type of dwelling.
\item floor5x: Floor level of the dwelling.
\item storeyx: Number of storeys in the dwelling.
\item number_of_rooms: Total number of rooms in the dwelling.
\item living.height: Height of the living room.
\item living.width: Width of the living room.
\item living.depth: Depth of the living room.
\item living.is_present: Indicator if the living room is present.
\item living.is_inspected: Indicator if the living room is inspected.
\item living.room_function: Function of the living room.
\item living.floor_level: Floor level of the living room.
\item kitchen.height: Height of the kitchen.
\item kitchen.width: Width of the kitchen.
\item kitchen.depth: Depth of the kitchen.
\item kitchen.is_present: Indicator if the kitchen is present.
\item kitchen.is_inspected: Indicator if the kitchen is inspected.
\item kitchen.room_function: Function of the kitchen.
\item kitchen.floor_level: Floor level of the kitchen.
\item bedroom.height: Height of the bedroom.
\item bedroom.width: Width of the bedroom.
\item bedroom.depth: Depth of the bedroom.
\item bedroom.is_present: Indicator if the bedroom is present.
\item bedroom.is_inspected: Indicator if the bedroom is inspected.
\item bedroom.room_function: Function of the bedroom.
\item bedroom.floor_level: Floor level of the bedroom.
\item bathroom.height: Height of the bathroom.
\item bathroom.is_present: Indicator if the bathroom is present.
\item bathroom.is_inspected: Indicator if the bathroom is inspected.
\item bathroom.floor_level: Floor level of the bathroom.
\item circulation.height: Height of the circulation area.
\item circulation.is_present: Indicator if the circulation area is present.
\item circulation.is_inspected: Indicator if the circulation area is inspected.
\item garage.is_present: Indicator if the garage is present.
\item garage.floor_level: Floor level of the garage.
\item balcony.is_present: Indicator if the balcony is present.
\item balcony.floor_level: Floor level of the balcony.
\item extra_rooms.is_present1: Indicator if the first extra room is present.
\item extra_rooms.is_present2: Indicator if the second extra room is present.
\item extra_rooms.is_present3: Indicator if the third extra room is present.
\item extra_rooms.is_present4: Indicator if the fourth extra room is present.
\item extra_rooms.is_present5: Indicator if the fifth extra room is present.
\item extra_rooms.is_present6: Indicator if the sixth extra room is present.
\item extra_rooms.is_present7: Indicator if the seventh extra room is present.
\item extra_rooms_1.room_function: Function of the first extra room.
\item extra_rooms_2.room_function: Function of the second extra room.
\item extra_rooms_3.room_function: Function of the third extra room.
\item extra_rooms_4.room_function: Function of the fourth extra room.
\item extra_rooms_5.room_function: Function of the fifth extra room.
\item extra_rooms_6.room_function: Function of the sixth extra room.
\item extra_rooms_7.room_function: Function of the seventh extra room.
\item extra_rooms_1.floor_level: Floor level of the first extra room.
\item extra_rooms_2.floor_level: Floor level of the second extra room.
\item extra_rooms_3.floor_level: Floor level of the third extra room.
\item extra_rooms_4.floor_level: Floor level of the fourth extra room.
\item extra_rooms_5.floor_level: Floor level of the fifth extra room.
\item extra_rooms_6.floor_level: Floor level of the sixth extra room.
\item extra_rooms_7.floor_level: Floor level of the seventh extra room.
\item conservatory.window_type: Type of windows in the conservatory.
\item conservatory.roof_type: Type of roof in the conservatory.
\item conservatory.closable_door: Indicator if the conservatory has a closable door.
\item conservatory.floor_area: Floor area of the conservatory.
\item stairs.stairs_within_dwelling: Indicator if there are stairs within the dwelling.
\item stairs.stairs_open_plan: Indicator if the stairs are open plan.
\item radiator.fixed_radiator: Indicator if there is a fixed radiator.
}
}
\usage{
s_ehs_2011_rom
}
\description{
This object contains the data retrieved from the English Housing Survey sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ehs_2011_rom
}
}
\seealso{
Other core data sources: 
\code{\link{s_bredem}},
\code{\link{s_chm}},
\code{\link{s_chm_reference}},
\code{\link{s_ehs_2011_ext}},
\code{\link{s_ehs_2011_hhd}},
\code{\link{s_sap}},
\code{\link{s_tus}}
}
\concept{core data sources}
\keyword{datasets}
