% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ehs.R
\docType{data}
\name{s_ehs_ml}
\alias{s_ehs_ml}
\title{Data from EHS}
\format{
A list containing various datasets related to the English Housing Survey (EHS) for machine learning purposes:
\itemize{
\item common: A tibble containing common data elements used across different datasets.
\item mlearning: A tibble containing data specifically prepared for machine learning applications.
\item merged: A tibble containing merged data from various sources for comprehensive analysis.
}
}
\usage{
s_ehs_ml
}
\description{
This object contains the data retrieved from the \strong{English Housing Survey} sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ehs_ml
}
}
\keyword{datasets}
