% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-metoffice.R
\docType{data}
\name{s_metoffice}
\alias{s_metoffice}
\title{Data from MetOffice}
\format{
A list with 1 data frame:
\itemize{
\item AvgExtTem: A tibble containing data related to average external temperature per month and GOR (Government Office Region)
}
}
\usage{
s_metoffice
}
\description{
This object contains the data retrieved from the MetOffice sources
}
\examples{
\dontrun{
# Displaying the whole list
s_metoffice
}
}
\keyword{datasets}
