% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-noabl.R
\docType{data}
\name{s_noabl}
\alias{s_noabl}
\title{Data from NOABL}
\format{
A list with 1 data frame:
\itemize{
\item AvgExtTem: A tibble containing data (Table 2011) related to average external wind speed per month and GOR (Government Office Region)
}
}
\usage{
s_noabl
}
\description{
This object contains the data retrieved from the Numerical Objective Analysis of Boundary Layer (NOABL) sources
}
\examples{
\dontrun{
# Displaying the whole list
s_noabl
}
}
\keyword{datasets}
