% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ons.R
\docType{data}
\name{s_ons}
\alias{s_ons}
\title{Data from ONS}
\format{
A list containing various datasets related to the Office for National Statistics (ONS) data.
\itemize{
\item lsoa.der: A tibble containing derived data for Lower Layer Super Output Areas (LSOA).
\item electricity: A tibble containing data on domestic electricity usage for LSOAs in 2017.
\item gas: A tibble containing data on domestic gas usage for LSOAs in 2017.
\item household: A tibble containing household data.
\item tenure: A tibble containing data on household tenure.
\item hhcomp: A tibble containing data on household composition.
}
}
\usage{
s_ons
}
\description{
This object contains the data retrieved from the Organisation for National Statistics (ONS) sources
}
\examples{
\dontrun{
# Displaying the whole list
s_ons
}
}
\keyword{datasets}
