% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tus.R
\docType{data}
\name{s_tus}
\alias{s_tus}
\title{Data from TUS}
\format{
A list containing various datasets related to the Time Use Survey (TUS).
\itemize{
\item paths: Data on paths taken by individuals.
\item tup_diary: Data from the time use diary.
\item tbl_individual: Data on individual characteristics.
\item tbl_household: Data on household characteristics.
\item tbl_diary: Data from the diary entries.
\item activities: Data on activities recorded in the survey.
\item activities_groups: Data on grouped activities.
\item age_filter: Data used for filtering by age.
\item essentials: Essential data elements for the survey.
\item lst_sessions: Data summarising activities per household
}
}
\usage{
s_tus
}
\description{
This object contains the data retrieved from the Time Use Survey (TUS) sources.
}
\examples{
\dontrun{
# Displaying the whole list
s_tus
}
}
\seealso{
Other core data sources: 
\code{\link{s_bredem}},
\code{\link{s_chm}},
\code{\link{s_chm_reference}},
\code{\link{s_ehs_2011_ext}},
\code{\link{s_ehs_2011_hhd}},
\code{\link{s_ehs_2011_rom}},
\code{\link{s_sap}}
}
\concept{core data sources}
\keyword{datasets}
