% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tus-ehs.R
\docType{data}
\name{s_tus_ehs}
\alias{s_tus_ehs}
\title{Data from TUS}
\format{
A list containing various datasets related to the Time Use Survey (TUS) and the English Housing Survey (EHS).
\itemize{
\item EHS: A tibble containing data from the English Housing Survey with 163,304 rows and 7 columns.
\itemize{
\item serial: Serial number of the survey entry.
\item .cross.income: Income category (low, mid, high).
\item .cross.tenure: Tenure type (Private, Social).
\item .cross.type: Dwelling type (Apartment, House).
\item .cross.region: Region of the dwelling.
\item .cross.employment: Employment status (irregular, paid employment, etc.).
\item .cross.hhtype: Household type (couple no dependants, etc.).
}
\item TUS: A data frame containing data from the Time Use Survey with 7,041 observations and 8 variables.
\itemize{
\item serial: Serial number of the survey entry.
\item .cross.income: Income category (low, mid, high).
\item .cross.tenure: Tenure type (Private, Social).
\item .cross.type: Dwelling type (Apartment, House).
\item .cross.region: Region of the dwelling.
\item .cross.employment: Employment status (irregular, paid employment, etc.).
\item .cross.hhtype: Household type (couple no dependants, etc.).
\item .cross.age: Age category (adult, child, etc.).
}
\item households: A tibble containing household data with 7,041 rows and 8 columns.
\itemize{
\item serial: Serial number of the survey entry.
\item .cross.income: Income category (high, mid, low).
\item .cross.tenure: Tenure type (Private, Social).
\item .cross.type: Dwelling type (House, Apartment).
\item .cross.region: Region of the dwelling.
\item .cross.employment: Employment status (paid employment, irregular, etc.).
\item .cross.hhtype: Household type (other type of household, couple no dependants, etc.).
\item .cross.age: Age category (adult, young, etc.).
}
}
}
\usage{
s_tus_ehs
}
\description{
This object contains the data retrieved from the Time Use Survey (TUS) and English Housing Survey (EHS) sources.
}
\examples{
\dontrun{
# Displaying the whole list
s_tus_ehs
}
}
\keyword{datasets}
