% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-uk-heff.R
\docType{data}
\name{s_uk_heff}
\alias{s_uk_heff}
\title{Data from UK-HEFF}
\format{
A list containing various datasets related to household energy efficiency in the UK.
\itemize{
\item population.raw: A tibble containing raw data on population.
\item households_x10e6.raw: A tibble containing raw data on households in millions.
\item households_x10e6.eng: A tibble containing data on households in millions for England.
\item tenure_x10e6.raw: A tibble containing raw data on tenure in millions.
\item type_x10e6.raw: A tibble containing raw data on dwelling types in millions.
\item fuelPrices.raw: A tibble containing raw data on fuel prices.
\item use24hrs: A tibble containing data on energy use over 24 hours.
\item seasonal: A tibble containing data on seasonal energy use.
\item power: A tibble containing data on power schedules.
\item schedules: A tibble containing data on various schedules.
\item typical: Data on typical energy use patterns.
\item inventory: Data on the inventory of energy-related items.
}
}
\usage{
s_uk_heff
}
\description{
This object contains the data retrieved from the UK-Housing Energy Fact File sources
}
\examples{
\dontrun{
# Displaying the whole list
s_uk_heff
}
}
\keyword{datasets}
