% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-worldbank.R
\docType{data}
\name{s_world_bank}
\alias{s_world_bank}
\title{Data from World Bank}
\format{
A list containing various datasets related to World Bank indicators.
\itemize{
\item d_carbon_intensity: A tibble containing data on carbon intensity with 59 rows and 11 columns.
\itemize{
\item year: Year of the data.
\item en.atm.ghgt.kt.ce: Total greenhouse gas emissions (kt of CO2 equivalent).
\item eg.use.pcap.kg.oe: Energy use (kg of oil equivalent per capita).
\item en.atm.co2e.kt: CO2 emissions (kt).
\item sp.pop.totl.ma.in: Population, male.
\item sp.pop.totl.fe.in: Population, female.
\item si.pov.gini: Gini index (measure of income inequality).
\item ny.gdp.mktp.kd: GDP (constant 2010 US$).
\item eg.use.elec.kh.pc: Electric power consumption (kWh per capita).
\item en.co2.bldg.zs: CO2 emissions from buildings (\% of total fuel combustion).
\item en.atm.co2e.eg.zs: CO2 emissions from electricity and heat production (\% of total fuel combustion).
}
\item lbl_carbon_intensity: A data frame containing metadata for the carbon intensity indicators with 10 rows and 3 columns.
\itemize{
\item Code: Indicator code.
\item Indicator.Name: Name of the indicator.
\item License.Type: License type for the data.
}
}
}
\usage{
s_world_bank
}
\description{
This object contains the data retrieved from the World Bank sources
}
\examples{
\dontrun{
# Displaying the whole list
s_world_bank
}
}
\keyword{datasets}
