% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplus-wrapper.R
\name{simulate_available_idfs}
\alias{simulate_available_idfs}
\title{Simulate available IDFs}
\usage{
simulate_available_idfs(sid)
}
\arguments{
\item{sid}{Simulation ID.}
}
\value{
Invisibly returns the input \code{sid}.
}
\description{
This function simulates available IDFs based on a
given simulation ID.
}
\seealso{
Other EnergyPlus evaluation: 
\code{\link{convert_idf_to_json}()},
\code{\link{extract_simulation_outputs}()},
\code{\link{generate_idf}()},
\code{\link{generate_idf_multiple}()},
\code{\link{read_idf_outcomes}()},
\code{\link{read_input_file}()},
\code{\link{read_input_summary}()},
\code{\link{run_energy_plus}()}
}
\concept{EnergyPlus evaluation}
