% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{summarise_heat_transfer_per_design_days}
\alias{summarise_heat_transfer_per_design_days}
\title{Summarizes heat load performance by design day}
\usage{
summarise_heat_transfer_per_design_days(l_hlcs, d_ddys)
}
\arguments{
\item{l_hlcs}{A list of heat load data frames.}

\item{d_ddys}{A data frame containing design day information.}
}
\value{
A list of lists, each containing summary data and design conditions for a specific design day.
}
\description{
This function summarizes heat load performance for each design day.
}
