% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{summarise_indexed_simulation}
\alias{summarise_indexed_simulation}
\title{Summarizes a simulation indexed by \code{i}}
\usage{
summarise_indexed_simulation(i)
}
\arguments{
\item{i}{Index of the simulation to process.}
}
\value{
A list containing:
\itemize{
\item \code{table}: A data frame with extracted simulation data.
\item \code{index}: A data frame with the index and UUID.
}
}
\description{
This function processes a simulation summary file, extracts relevant data,
and returns it in a structured format.
}
