% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{summarise_zone_dimensions}
\alias{summarise_zone_dimensions}
\title{Summarizes Zone Dimensions}
\usage{
summarise_zone_dimensions(do, dz)
}
\arguments{
\item{do}{A data frame containing surface information.}

\item{dz}{A data frame containing zone information.}
}
\value{
A list containing zone-specific information:
\itemize{
\item \code{surfaces}: Surface data for the zone.
\item \code{zone}: Zone-level data for the zone.
}
}
\description{
Summarizes zone dimensions and surface information for each thermal zone.
}
