% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--layouts.R
\docType{data}
\name{t_layouts}
\alias{t_layouts}
\title{A tupple with room/layout information}
\format{
A list of 50 elements, each representing a different layout. Each layout is a list with the following components:
\describe{
\item{loc}{\code{<int>}: A vector of integers representing the locations of the rooms within the layout.}
\item{size}{\code{<list>}: A list containing the dimensions of the layout:
\describe{
\item{cols}{\code{<int>}: Number of columns in the layout.}
\item{rows}{\code{<int>}: Number of rows in the layout.}
}
}
}
The following example shows the structure of layout no. 069:

\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
  "loc": [9, 61, 23, 27],
  "size": \{
      "cols": [2],
      "rows": [2]
    \},
  "shape": ["R"],
  "proportion": [
      \{
        "name": "Stairwell",
        "pct": 0.25
      \},
      \{
        "name": "Bathroom",
        "pct": 0.25
      \},
      \{
        "name": "Bedroom Single",
        "pct": 0.25
      \},
      \{
        "name": "Bedroom Single (B)",
        "pct": 0.25
      \}
    ]
\}
}\if{html}{\out{</div>}}
}
\usage{
t_layouts
}
\description{
This tupple contains information about various room layouts, including their locations, sizes, shapes, and proportions of different room types.
}
\examples{
\dontrun{
# Displaying the whole list
t_layouts

# Displaying the layout 069
t_layouts$lay_069 \%>\% jsonlite::toJSON(pretty = TRUE)
}
}
\seealso{
Other room layouts: 
\code{\link{d_rooms}}
}
\concept{room layouts}
\keyword{datasets}
