% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{toggle_chronometer}
\alias{toggle_chronometer}
\title{Toggles a Chronometer}
\usage{
toggle_chronometer()
}
\description{
This function starts or stops a chronometer to
measure elapsed time.
}
\details{
The function uses a global variable \code{.varChrono} to store the start time.
If \code{.varChrono} exists, it calculates and displays the elapsed time
and then removes the variable. Otherwise, it sets the start time to
the current \code{proc.time()}.
}
