% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.R
\name{update_hlp_layers}
\alias{update_hlp_layers}
\title{Updates layer properties in a surface object}
\usage{
update_hlp_layers(l_src, d_upd)
}
\arguments{
\item{l_src}{A list representing the surface object, containing layers and thermal properties.}

\item{d_upd}{A data frame containing updated layer properties.}
}
\value{
An updated surface object with the replaced layer and recalculated properties.
}
\description{
Replaces the first layer of a surface object with updated properties and recalculates thermal resistance and transmittance.
}
