% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typology.R
\name{use_custom_file_subset}
\alias{use_custom_file_subset}
\title{Subset EHS data based on an external file with user-defined criteria}
\usage{
use_custom_file_subset(
  file_custom = system.file("extdata", "custom_input.dcf", package = "enhub"),
  d_ehs = combine_ehs(),
  k_input = get_input_set()
)
}
\arguments{
\item{file_custom}{A file containing user-defined criteria for subsetting EHS data. A default file is provided.}

\item{d_ehs}{A data frame containing EHS data (assumed to be combined from multiple sources).}

\item{k_input}{A list containing initial indices for subsetting (optional, defaults to all rows).}
}
\value{
A data frame containing the subset of EHS data based on user-defined criteria.
}
\description{
This function allows users to subset EHS data based on user-defined
criteria stored in an external file (e.g., \code{custom_input.dcf}, \code{custom_input.json}).

The file should contain a list of variables and their corresponding values for
subsetting the data. The purpose of this approach is to allow users to define
custom subsets without modifying the code, with that definition stored in a
separate file, potentially in the cloud.
}
\details{
The file specifies various attributes related to dwellings, regions, tenure types,
and other survey variables, for example:

\describe{
\item{D002_DwellingType}{detached, semi detached, mid terrace, end terrace, purpose built, converted, non residential}
\item{D003_Region}{does not apply, no answer, North East, Yorkshire and the Humber, North West, East Midlands, West Midlands, South West, Eastern England, South East, London}
\item{D004_TenureType}{housing association, local authority, private owner occupied, private rented}
\item{V573_Urbanity}{City centre, Urban, Suburban residential, Rural residential, Village centre, Rural, Question Not Applicable, Unknown}
\item{V542_FloorPosition}{House/Bungalow, Top Floor Flat, Mid Floor Flat, Ground floor flat, Basement Flat, Unknown}
\item{V534_TFAsurvey}{<50, 200-250, >500}
\item{D001_DwellingAge}{>1,< 11}
\item{D082_MainHeatingSystemType}{1-15}
\item{D097_DHWSystemType}{1-11}
\item{V576_CuboidAttic}{TRUE, FALSE}
\item{D030_WallThickness}{0.15-0.64}
\item{V563_GlazingRatio}{0-1}
\item{D109_SolarDHWSystem}{TRUE, FALSE}
}

An example of the provided template (retrievable via \code{system.file("extdata", "custom_input.dcf", package = "enhub")}) contains the following entries:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{D002_DwellingType: detached, semi detached, mid terrace, end terrace
D003_Region: Yorkshire and the Humber, North East
D004_TenureType: private owner occupied
V542_FloorPosition: House/Bungalow
V534_TFAsurvey: >150
D001_DwellingAge: < 10
D082_MainHeatingSystemType: 1-15
D097_DHWSystemType: 1-11
V576_CuboidAttic: TRUE
}\if{html}{\out{</div>}}
}
